% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages.R
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame.gplin}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{gplin}(x, ...)
}
\arguments{
\item{x}{any \R object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Functions to coerce an object to a data frame.
}
\details{
An object of class \sQuote{gplin} is coerced to a data frame
by means of \code{as.data.frame}.
}
\keyword{internal}
