% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{batch}
\alias{batch}
\title{batch}
\usage{
batch(object)
}
\arguments{
\item{object}{either the name of the model to end with batch sampling
\emph{or} a previously computed \sQuote{gpsim} object}
}
\value{
An object of \sQuote{gpsim} class.
}
\description{
Batch sampling at the end of process
}
\seealso{
More example genealogy processes: 
\code{\link{lbdpwr}},
\code{\link{lbdp}},
\code{\link{moranwr}},
\code{\link{moran}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sirwr}},
\code{\link{sir}}
}
\concept{Genealogy processes}
