% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdei.R, R/bdei_pomp.R
\name{bdei}
\alias{bdei}
\alias{BDEI}
\alias{runBDEI}
\alias{continueBDEI}
\alias{bdei_pomp}
\title{Linear birth-death with exposed and infectious classes}
\usage{
runBDEI(
  time,
  t0 = 0,
  sigma = 1/7,
  lambda = 7/9,
  mu = 1/6,
  chi = 1/6,
  pop = 1,
  E0 = 0,
  I0 = 1
)

continueBDEI(object, time, sigma = NA, lambda = NA, mu = NA, chi = NA)

bdei_pomp(x, sigma, lambda, mu, chi, E0, I0, pop)
}
\arguments{
\item{time}{end timepoint of simulation}

\item{t0}{beginning timepoint of simulation}

\item{sigma}{per capita progression rate (exposed to infectious)}

\item{lambda}{per capita reproduction rate}

\item{mu}{per capita removal rate (of infectious lineages)}

\item{chi}{(destructive) sampling rate}

\item{pop}{initial population size}

\item{E0}{initial fraction of exposed lineages}

\item{I0}{initial fraction of infectious lineages}

\item{object}{a previously computed simulation}

\item{x}{genealogy in \pkg{phylopomp} format.}
}
\value{
\code{runBDEI} and \code{continueBDEI} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{BDEI}.

\code{bdei_pomp} returns a \sQuote{pomp} object.
}
\description{
Two-deme linear birth-death process with exposed (incubating) and infectious lineages.
Transmission creates exposed offspring from infectious donors; exposed lineages progress to infectious.
Only infectious lineages are subject to sampling and sampling is destructive.
Identical to the BDEI model of Voznica et al. (2022, https://doi.org/10.1038/s41467-022-31511-0), albeit with a different parameterization.
}
\details{
\code{bdei_pomp} constructs a \sQuote{pomp} object containing a given set of data and a BDEI model.
}
\examples{
library(phylopomp)

runBDEI(time=10, pop=2, E0=0.5, I0=0.5) |>
  plot()
}
\references{
\Voznica2022
}
\seealso{
More example genealogy processes:
\code{\link{bdss}},
\code{\link{lbdp}},
\code{\link{mers}},
\code{\link{moran}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link[=simulate]{simulate()}},
\code{\link{sir}},
\code{\link{strains}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
