% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continue.R
\name{continue}
\alias{continue}
\alias{continue.gpsim}
\title{Continue}
\usage{
continue(data, ...)

\method{continue}{gpsim}(data, times, ...)
}
\arguments{
\item{data}{previously computed \sQuote{gpsim} object.}

\item{...}{additional arguments, for example to change model parameters.}

\item{times}{times at which output is requested.}
}
\value{
A \code{tibble} with \code{state} attribute.
}
\description{
Continue previous run
}
\section{Note}{

One cannot change initial conditions or \code{t0} with \code{continue}.
}

\seealso{
Other Genealogy processes: 
\code{\link{siir}},
\code{\link{sir}}
}
\concept{Genealogy processes}
