% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curtail.R
\name{curtail}
\alias{curtail}
\title{Curtail a genealogy to the given time}
\usage{
curtail(object, time = NA, t0 = NA)
}
\arguments{
\item{object}{\code{gpsim} object.}

\item{t0, time}{numeric scalars; determine the time interval for curtailed genealogy}
}
\value{
a genealogy object: the curtailed version of the input genealogy.
}
\description{
Discards all nodes beyond the given time.
}
\examples{
library(ggplot2)

simulate("SIIR",time=5) -> x

plot_grid(
  x |>
    plot(prune=FALSE,points=TRUE),
  x |>
    curtail(time=3) |>
    plot(prune=FALSE,points=TRUE)+
    expand_limits(x=5),
  ncol=1,align="h",axis="tblr"
)

plot_grid(
  x |>
    plot(prune=TRUE,points=TRUE)+
    geom_vline(xintercept=3),
  x |> curtail(time=3) |>
    plot(prune=TRUE,points=TRUE)+
    geom_vline(xintercept=3)+
    expand_limits(x=5),
  ncol=1,align="h",axis="tblr"
)
}
