% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{diagram}
\alias{diagram}
\title{Genealogy process diagram}
\usage{
diagram(data, ..., vp = NULL)
}
\arguments{
\item{data}{\code{gpsim} object.}

\item{...}{graphical parameter settings, suitable for passing to \code{\link[grid:gpar]{gpar}}.}

\item{vp}{a \code{\link[grid]{viewport}} object (or \code{NULL}).}
}
\value{
A \pkg{grid} graphics object (\code{grob}), invisibly.
}
\description{
Produces a diagram of the genealogy process state.
}
\examples{
library(cowplot)
playSIR(Beta=3,gamma=0.1,psi=0.2,S0=100,I0=5,R0=0,times=2,t0=0) -> x
y <- getInfo(x,prune=FALSE,compact=FALSE)
plot(y,points=TRUE)
y <- getInfo(x)
plot_grid(plotlist=list(plot(y,points=TRUE)[[1]],diagram(y)),
  ncol=1,rel_heights=c(2,1))

plot_grid(diagram(y))
}
