% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneal.R
\name{geneal}
\alias{geneal}
\title{Bare genealogy}
\usage{
geneal(object)
}
\arguments{
\item{object}{a \sQuote{gpsim} object.}
}
\value{
A bare genealogy object.
}
\description{
Extracts the bare genealogy from a Markov genealogy process simulation
}
