% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{geneal_scale}
\alias{geneal_scale}
\title{Rescale and/or shift time.}
\usage{
geneal_scale(x, scale, origin = 0)
}
\arguments{
\item{x}{genealogy to be scaled}

\item{scale, origin}{scalars}
}
\value{
modified genealogy
}
\description{
Time-rescaling and/or shifting.
}
\details{
A linear transformation is applied to the genealogy's time-scale.
Specifically, if \eqn{t} is the old time and \eqn{t'} the new time,
then \eqn{t' = a (t - b)} where \eqn{a=}\code{scale} and
\eqn{b=}\code{origin}.
}
\examples{
library(ggplot2)

runSEIR(t0=0,time=3,pop=200,E0=0.02,I0=0) -> x
geneal_scale(x,365,3) -> y
plot_grid(plot(x),plot(y),ncol=1)
}
