% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{geneal_sum}
\alias{geneal_sum}
\alias{+.gpgen}
\title{Combine genealogies}
\usage{
\method{+}{gpgen}(x, y)

geneal_sum(...)
}
\arguments{
\item{x, y, ...}{genealogies or lists of genealogies to be summed.}
}
\value{
combined genealogy.
}
\description{
A summing operator for genealogies.
}
\details{
The sum of genealogies is defined to be the union of all the lineages,
restricted to the common time-interval, i.e., the intersection of all the individual time-intervals.
}
\examples{
library(ggplot2)

runSEIR(t0=-1,time=2,E0=0,I0=0.03) -> x
runSEIR(t0=0,time=3,E0=0.02,I0=0) -> y
runSEIR(t0=1,time=4,E0=0.03,I0=0.02) -> z

plot_grid(
  plot_grid(
    plotlist=lapply(
      list(x,y,z),
      plot,
      points=TRUE,ladderize=FALSE,
      legend.position="none"
    ) |>
      lapply(\(.).+geom_vline(xintercept=c(-1,0,1,2,3,4))),
    labels=c("x","y","z"),
    ncol=1,align="v",axis="bltr"
  ),
  plot_grid(
    plotlist=lapply(
      list(x+y,y+z,x+y+z),
      plot,
      points=TRUE,ladderize=FALSE,
      legend.position="none"
    ) |>
      lapply(\(.).+geom_vline(xintercept=c(-1,0,1,2,3,4))),
    labels=c("x+y","y+z","x+y+z"),
    ncol=1,align="h",axis="bltr"
  )+
    geom_vline(xintercept=c(0,1,2,3,4)),
  nrow=1
)
}
