% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getinfo.R
\name{getInfo}
\alias{getInfo}
\title{getInfo}
\usage{
getInfo(
  object,
  prune = TRUE,
  obscure = TRUE,
  extended = TRUE,
  t0 = FALSE,
  time = FALSE,
  structure = FALSE,
  yaml = FALSE,
  ndeme = FALSE,
  lineages = FALSE,
  newick = FALSE,
  nsample = FALSE,
  nroot = FALSE,
  genealogy = FALSE,
  gendat = FALSE,
  cblv = FALSE
)
}
\arguments{
\item{object}{\code{gpsim} object.}

\item{prune}{logical; prune the genealogy?}

\item{obscure}{logical; obscure the demes?}

\item{extended}{logical; return extended-Newick format?}

\item{t0}{logical; return the zero-time?}

\item{time}{logical; return the current time?}

\item{structure}{logical; return the structure in \R list format?}

\item{yaml}{logical; return the structure in YAML format?}

\item{ndeme}{logical; return the number of demes?}

\item{lineages}{logical; return the lineage-count function?}

\item{newick}{logical; return a Newick-format description of the tree?}

\item{nsample}{logical; return the number of samples?}

\item{nroot}{logical; return the number of roots?}

\item{genealogy}{logical; return the lineage-traced genealogy?}

\item{gendat}{logical; return the data-frame format?}

\item{cblv}{logical; return the cblv representation?}
}
\value{
A list containing the requested elements, including any or all of:
\describe{
  \item{t0}{the initial time (a numeric scalar)}
  \item{time}{the final time (a numeric scalar)}
  \item{ndeme}{the number of demes (an integer)}
  \item{nsample}{the number of samples (an integer)}
  \item{nroot}{the number of roots (an integer)}
  \item{newick}{the genealogical tree, in Newick format (extended-Newick if \code{extended=TRUE})}
  \item{yaml}{the state of the genealogy process in YAML format}
  \item{structure}{the state of the genealogy process in \R list format}
  \item{lineages}{a \code{\link[tibble]{tibble}} containing the lineage count function through time}
  \item{gendat}{a \code{\link[tibble]{tibble}} containing the (obscured) genealogy in a data-frame format}
  \item{genealogy}{the lineage-traced genealogy (as a raw vector)}
  \item{cblv}{the cblv representation (as a matrix)}
}
}
\description{
Retrieve information from genealogy process simulation
}
\examples{
simulate("SIIR",time=3,psi1=1,psi2=0) |>
  simulate(Beta1=2,gamma=2,time=10,psi1=10,psi2=1) |>
  plot()

runSIIR(Beta1=10,Beta2=8,
  pop=220,S_0=200,I1_0=10,I2_0=8,R_0=0,time=0,t0=-1) |>
  simulate(psi1=10,time=2) |>
  plot(points=TRUE,obscure=FALSE)

simulate("SIIR",Beta1=2,Beta2=50,gamma=1,psi1=2,
  pop=322,S_0=300,I1_0=20,I2_0=2,time=5) |>
  lineages() |>
  plot()
}
\keyword{internal}
