% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{genealogyGrob}
\alias{genealogyGrob}
\alias{nodeGrob}
\alias{pocketGrob}
\alias{ballGrob}
\alias{resizingTextGrob}
\alias{drawDetails.resizingTextGrob}
\alias{preDrawDetails.resizingTextGrob}
\alias{postDrawDetails.resizingTextGrob}
\alias{drawDetails.ballGrob}
\alias{preDrawDetails.ballGrob}
\alias{postDrawDetails.ballGrob}
\title{Diagramming internals}
\usage{
genealogyGrob(data, n = length(data$nodes), vp = NULL)

nodeGrob(data, vp = NULL)

pocketGrob(data, vp = NULL)

ballGrob(data, vp = NULL)

resizingTextGrob(..., vp = NULL)

\method{drawDetails}{resizingTextGrob}(x, recording = TRUE)

\method{preDrawDetails}{resizingTextGrob}(x)

\method{postDrawDetails}{resizingTextGrob}(x)

\method{drawDetails}{ballGrob}(x, recording = TRUE)

\method{preDrawDetails}{ballGrob}(x)

\method{postDrawDetails}{ballGrob}(x)
}
\arguments{
\item{data}{list; pocket structure}

\item{n}{length of longest genealogy}

\item{vp}{a \code{\link[grid]{viewport}} object (or \code{NULL}).}

\item{...}{arguments to be passed to \code{\link[grid:textGrob]{textGrob}}.}

\item{x}{ An \R object. }

\item{recording}{ A logical value indicating whether a grob
  is being added to the display list or redrawn from the display list. }
}
\description{
Facilities to produce diagrammatic representations
of genealogy process states.
}
\details{
Code for the resizing text adapted from a blog post by Mark Heckmann
(https://ryouready.wordpress.com/2012/08/01/creating-a-text-grob-that-automatically-adjusts-to-viewport-size/).
}
\keyword{internal}
