% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leventhal.R
\name{leventhal}
\alias{leventhal}
\alias{leventhal_pomp}
\title{Leventhal (2014) SI model}
\usage{
leventhal_pomp(data, beta, gamma, psi, N, I0 = 1, t0 = 0)
}
\arguments{
\item{data}{optional data frame; output from \code{playLeventhal}.}

\item{beta}{contact rate}

\item{gamma}{recovery rate}

\item{psi}{sampling rate}

\item{N}{population size}

\item{I0}{initial number of infections}

\item{t0}{initial time}
}
\description{
Simulation and inference for Leventhal's model
}
\details{
\code{leventhal_pomp} constructs a \pkg{pomp} object containing a given set of data and a Leventhal-type SIS infection.

It is assumed that \code{data} is in the format returned by \code{\link{newick2df}}.
}
\seealso{
Other Genealogy processes: 
\code{\link{lbdp}},
\code{\link{moran}},
\code{\link{sirs}},
\code{\link{sirws}}
}
\concept{Genealogy processes}
