% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mers.R
\name{mers}
\alias{mers}
\alias{MERS}
\alias{runMERS}
\alias{continueMERS}
\title{Two-host infection model with spillover and demography.
Hosts are culled upon sampling.}
\usage{
runMERS(
  time,
  t0 = 0,
  Beta_cc = 4,
  Beta_ch = 0,
  Beta_hc = 0,
  Beta_hh = 4,
  gamma_c = 1,
  gamma_h = 1,
  chi_c = 1,
  chi_h = 0,
  Bc = 0.1,
  Bh = 0.03,
  Sc0 = 1,
  Sh0 = 1,
  Ic0 = 0.01,
  Ih0 = 0,
  Nc = 10000,
  Nh = 10000
)

continueMERS(
  object,
  time,
  Beta_cc = NA,
  Beta_ch = NA,
  Beta_hc = NA,
  Beta_hh = NA,
  gamma_c = NA,
  gamma_h = NA,
  chi_c = NA,
  chi_h = NA,
  Bc = NA,
  Bh = NA
)
}
\arguments{
\item{time}{end timepoint of simulation}

\item{t0}{beginning timepoint of simulation}

\item{Beta_cc}{transmission rate within camels}

\item{Beta_ch}{transmission from human to camel}

\item{Beta_hc}{transmission from camel to human}

\item{Beta_hh}{transmission rate within humans}

\item{gamma_c}{removal rate in camels}

\item{gamma_h}{removal rate in humans}

\item{chi_c}{per capita destructive sampling rate for camels}

\item{chi_h}{per capita destructive sampling rate for humans}

\item{Bc}{gross birth rate for camels}

\item{Bh}{gross birth rate for humans}

\item{Sc0}{initial susceptible fraction of camel population}

\item{Sh0}{initial susceptible fraction of human population}

\item{Ic0}{initial infected fraction of camel population}

\item{Ih0}{initial infected fraction of human population}

\item{Nc}{camel population size}

\item{Nh}{human population size}

\item{object}{a previously computed simulation}
}
\value{
\code{runMERS} and \code{continueMERS} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{MERS}.
}
\description{
The population is structured by infection progression
and host species.
}
\seealso{
More example genealogy processes:
\code{\link{bdei}},
\code{\link{bdss}},
\code{\link{lbdp}},
\code{\link{moran}},
\code{\link{mtbd2}},
\code{\link{mtbd3}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link[=simulate]{simulate()}},
\code{\link{sir}},
\code{\link{strains}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
