% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranwr.R
\name{moranwr}
\alias{moranwr}
\alias{Moranwr}
\alias{runMoranwr}
\alias{continueMoranwr}
\title{The classical Moran model with reassortment}
\usage{
runMoranwr(
  time,
  t0 = 0,
  n = 100,
  mu = 1,
  psi = 1,
  rhoA = 0,
  rhoB = 0,
  frac = 0
)

continueMoranwr(
  object,
  time,
  mu = NA,
  psi = NA,
  rhoA = NA,
  rhoB = NA,
  frac = NA
)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{n}{population size}

\item{mu}{event rate}

\item{psi}{sampling rate}

\item{rhoA}{reassortment rate for segment A}

\item{rhoB}{reassortment rate for segment B}

\item{frac}{fraction of batch sampling}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}
}
\value{
An object of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{Moran}.
}
\description{
The Markov genealogy process induced by the classical Moran process,
in which birth/death events occur at a constant rate and the
population size remains constant.
}
\seealso{
More example genealogy processes: 
\code{\link{batch}()},
\code{\link{lbdpwr}},
\code{\link{lbdp}},
\code{\link{moran}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sirwr}},
\code{\link{sir}}
}
\concept{Genealogy processes}
