% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtbd3.R
\name{mtbd3}
\alias{mtbd3}
\alias{MTBD3}
\alias{runMTBD3}
\alias{continueMTBD3}
\title{General 3-type linear multitype birth-death-sampling process (BDMM, Stadler psi/r parameterization)}
\usage{
runMTBD3(
  time,
  t0 = 0,
  lambda_1_1 = 1,
  lambda_1_2 = 0.1,
  lambda_1_3 = 0.1,
  lambda_2_1 = 0.1,
  lambda_2_2 = 1,
  lambda_2_3 = 0.1,
  lambda_3_1 = 0.1,
  lambda_3_2 = 0.1,
  lambda_3_3 = 1,
  m_1_2 = 0,
  m_1_3 = 0,
  m_2_1 = 0,
  m_2_3 = 0,
  m_3_1 = 0,
  m_3_2 = 0,
  mu_1 = 0.5,
  mu_2 = 0.5,
  mu_3 = 0.5,
  psi_1 = 0.1,
  psi_2 = 0.1,
  psi_3 = 0.1,
  r_1 = 1,
  r_2 = 1,
  r_3 = 1,
  I1_0 = 1,
  I2_0 = 0,
  I3_0 = 0
)

continueMTBD3(
  object,
  time,
  lambda_1_1 = NA,
  lambda_1_2 = NA,
  lambda_1_3 = NA,
  lambda_2_1 = NA,
  lambda_2_2 = NA,
  lambda_2_3 = NA,
  lambda_3_1 = NA,
  lambda_3_2 = NA,
  lambda_3_3 = NA,
  m_1_2 = NA,
  m_1_3 = NA,
  m_2_1 = NA,
  m_2_3 = NA,
  m_3_1 = NA,
  m_3_2 = NA,
  mu_1 = NA,
  mu_2 = NA,
  mu_3 = NA,
  psi_1 = NA,
  psi_2 = NA,
  psi_3 = NA,
  r_1 = NA,
  r_2 = NA,
  r_3 = NA
)
}
\arguments{
\item{time}{end timepoint of simulation}

\item{t0}{beginning timepoint of simulation}

\item{lambda_1_1}{birth rate type1 parent -> type1 child}

\item{lambda_1_2}{birth rate type1 parent -> type2 child}

\item{lambda_1_3}{birth rate type1 parent -> type3 child}

\item{lambda_2_1}{birth rate type2 parent -> type1 child}

\item{lambda_2_2}{birth rate type2 parent -> type2 child}

\item{lambda_2_3}{birth rate type2 parent -> type3 child}

\item{lambda_3_1}{birth rate type3 parent -> type1 child}

\item{lambda_3_2}{birth rate type3 parent -> type2 child}

\item{lambda_3_3}{birth rate type3 parent -> type3 child}

\item{m_1_2}{migration type1 -> type2}

\item{m_1_3}{migration type1 -> type3}

\item{m_2_1}{migration type2 -> type1}

\item{m_2_3}{migration type2 -> type3}

\item{m_3_1}{migration type3 -> type1}

\item{m_3_2}{migration type3 -> type2}

\item{mu_1}{unobserved removal rate in type1}

\item{mu_2}{unobserved removal rate in type2}

\item{mu_3}{unobserved removal rate in type3}

\item{psi_1}{overall sampling rate for type1}

\item{psi_2}{overall sampling rate for type2}

\item{psi_3}{overall sampling rate for type3}

\item{r_1}{probability a sampled type1 individual is removed}

\item{r_2}{probability a sampled type2 individual is removed}

\item{r_3}{probability a sampled type3 individual is removed}

\item{I1_0}{initial lineages in type1}

\item{I2_0}{initial lineages in type2}

\item{I3_0}{initial lineages in type3}

\item{object}{a previously computed simulation}
}
\value{
\code{runMTBD3} and \code{continueMTBD3} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{MTBD3}.
}
\description{
A linear multitype birth-death-sampling process on 3 demes (type1..type3),
following the continuous reaction set of Vaughan & Stadler (2025, MBE).
Per capita, a type i individual: gives birth to a type j offspring at rate
lambda_i_j (parent stays type i); migrates i->j at rate m_i_j; is removed
unobserved at rate mu_i; and is sampled at overall rate psi_i, being removed
on sampling with probability r_i (SampR, sample_death) or retained with
probability 1-r_i (SampNR, sampled ancestor). Present-time (rho) sampling and
skyline time-variation are NOT included. Generated by gen_bdmm.R -- do not
hand-edit; regenerate.
}
\seealso{
More example genealogy processes:
\code{\link{bdei}},
\code{\link{bdss}},
\code{\link{lbdp}},
\code{\link{mers}},
\code{\link{moran}},
\code{\link{mtbd2}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link[=simulate]{simulate()}},
\code{\link{sir}},
\code{\link{strains}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
