% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisirws.R
\name{multisirws}
\alias{multisirws}
\alias{playmultiSIRwS}
\alias{multiSIRwS}
\alias{multisir_pomp}
\title{Overdispered SIR with sampling simulator.}
\usage{
playmultiSIRwS(
  data = NULL,
  ...,
  beta,
  gamma,
  psi,
  theta = Inf,
  S0,
  I0,
  R0,
  t0 = 0,
  times,
  tree = FALSE,
  ill = FALSE
)

multisir_pomp(data, beta, gamma, psi, theta, S0, I0, R0, N, t0 = 0)
}
\arguments{
\item{data}{optional data frame; output from \code{playmultiSIRwS}.}

\item{...}{additional arguments; ignored.}

\item{beta}{transmission rate}

\item{gamma}{recovery rate}

\item{psi}{sampling rate}

\item{theta}{inverse over-dispersed parameter}

\item{S0}{initial size of susceptible population}

\item{I0}{initial size of infected population}

\item{R0}{initial size of recovered population}

\item{t0}{initial time}

\item{times}{times at which output is requested.}

\item{tree}{logical; represent the genealogical tree in Newick format?}

\item{ill}{logical; return an illustration?}

\item{N}{initial total population size}
}
\value{
A \code{tibble} with \code{state} attribute.
}
\description{
Run the simulator.
}
\details{
\code{multisir_pomp} constructs a \pkg{pomp} object containing a given set of data and an overdispersed SIR model.

It is assumed that \code{data} is in the format returned by \code{\link{nwk2df}}.
}
\examples{
playmultiSIRwS(beta=10,gamma=1,psi=.5,theta=100.5,S0=500,I0=5,R0=0,times=0:5,t0=-1,tree=TRUE) -> x
plot(x,points=TRUE)

y <- getInfo(x,compact=TRUE)
plot(y,points=TRUE)

library(ggplot2)
y$lineages \%>\%
  ggplot(aes(x=time,y=lineages))+
  geom_step()


playmultiSIRwS(beta=5,gamma=1,psi=.5,theta=100.5,S0=500,I0=5,R0=0,times=5,t0=0,tree=TRUE) \%>\%
  getInfo(compact=TRUE) -> x

plot(x,points=TRUE)

library(pomp)

x$tree \%>\%
  nwk2df(time=5) \%>\%
  multisir_pomp(beta=5,gamma=1,psi=.5,theta=100.5,S0=500,I0=5,R0=0,N=505) \%>\%
  pfilter(Np=2000) \%>\%
  logLik()
}
\seealso{
Other Genealogy processes: 
\code{\link{lbdp}},
\code{\link{leventhal}},
\code{\link{moran}},
\code{\link{sirws}}
}
\concept{Genealogy processes}
