% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwk2df.R
\name{nwk2df}
\alias{nwk2df}
\title{Convert a tree in Newick format to data frame}
\usage{
nwk2df(tree, time = NA, root_time = 0, show_branches = FALSE)
}
\arguments{
\item{tree}{tree data in Newick format.}

\item{time}{time of the genealogy.}

\item{root_time}{time of the root.}

\item{show_branches}{whether to show number of new branches at coalesecens.}
}
\value{
A data frame suitable for use as \code{pomp} input, containing three columns:
\describe{
  \item{time}{numeric; time of the genealogy event.}
  \item{lineages}{integer; the value of the lineage-count function at the specified time.
    Note that this function is right-continuous with left limits, and constant on the inter-event intervals.
  }
  \item{code}{integer; a code describing the nature of the event.
    1 indicates a coalescence;
    0 indicates a dead sample;
   -1 indicates a leaf sample;
    2 indicates the root;
    9 indicates the end time of sampling.
  }
}
}
\description{
Convert a genealogical tree in Newick format to a data frame suitable for use with \pkg{pomp}.
}
\details{
If \code{time} is furnished, it is assumed that the absence of samples between the latest leaf and \code{time} is informative.

Invisible nodes (labeled 'X_' for any X) are dropped.
}
\examples{
playSIRwS(beta=2,gamma=1,psi=2,S0=100,I0=2,times=c(0,5),t0=0,tree=TRUE) -> x
y <- getInfo(x)
nwk2df(y$tree, time=5, show_branches=TRUE) -> z

# compare to y$lineages
y$lineages \%>\%
  all.equal(tail(z,-1)[,1:2],tolerance=1e-5)
}
