% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_newick}
\alias{parse_newick}
\title{parse a Newick string}
\usage{
parse_newick(x, t0 = 0, tf = NA)
}
\arguments{
\item{x}{character; the Newick description.
See Details for specifics.}

\item{t0}{numeric; the root time.}

\item{tf}{numeric; the current or final time.}
}
\value{
An object of class \dQuote{gpgen}.
}
\description{
Parses a Newick description and returns a binary version of the genealogy.
}
\details{
\code{parse_newick} parses a string containing information in an extended Newick format and returns a \pkg{phylopomp} genealogy.

The extension allows metadata (enclosed by square brackets).
It ignores metadata unless they are specifically flagged as PhyloPOMP metadata, i.e.:
\preformatted{
   [&&PhyloPOMP:...]
}
PhyloPOMP metadata can be used to encode the node type (i.e., \code{type=sample}, \code{type=extant}, \code{type=node}) and/or the deme (given as an integer, e.g., \code{deme=1}).
}
