% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{phylopomp}
\alias{phylopomp}
\alias{phylopomp,package}
\alias{phylopomp-package}
\title{Phylodynamics for POMP models}
\description{
Simulation and inference of Markov genealogy processes.
}
\author{
Aaron A. King, Qianying Lin
}
