% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.gpgen}
\title{Fancy tree plotter}
\usage{
\method{plot}{gpgen}(
  x,
  ...,
  prune = TRUE,
  obscure = TRUE,
  points = FALSE,
  ladderize = TRUE,
  palette = scales::hue_pal(l = 30, h = c(220, 580))
)
}
\arguments{
\item{x}{object of class \sQuote{gpgen}}

\item{...}{\code{plot} passes extra arguments to \code{\link[ggplot2]{theme}}.}

\item{prune}{logical; prune the genealogy?}

\item{obscure}{logical; obscure the demes?}

\item{points}{logical; show nodes and tips?}

\item{ladderize}{logical; ladderize?}

\item{palette}{color palette for indicating demes.
This can be furnished either as a function or a vector of colors.
If this is a function, it should take a single integer argument, the number of colors required.
If it is a vector, it should have at least as many elements as there are demes in the genealogy.}
}
\value{
A printable \code{ggplot} object.
}
\description{
Plots a genealogical tree.
}
\details{
Tree-plotting in \pkg{phylopomp} depends on the \pkg{ggtree} package.
}
\examples{
simulate("SEIR",Beta=2,sigma=2,gamma=1,psi=2,S0=1000,I0=5,time=5) |>
  simulate(Beta=5,gamma=2,time=10,psi=3) |>
  plot()

runSEIR(Beta=3,gamma=1,psi=2,S0=20,I0=5,R0=0,time=5,t0=-1) |>
  plot(points=TRUE,obscure=FALSE)

runSEIR(Beta=3,gamma=0.1,psi=0.2,S0=100,I0=5,R0=0,time=2,t0=0) -> x
plot_grid(plotlist=list(plot(x,points=TRUE),diagram(x)),
  ncol=1,rel_heights=c(4,1))

simulate("SEIR",sigma=1,omega=1,time=20,I0=4) |> plot(obscure=FALSE)

simulate("SEIR",sigma=1,omega=1,time=20,I0=4) |>
  lineages(obscure=FALSE) |>
  plot()
simulate("LBDP",time=4) |> plot(points=TRUE)

simulate("LBDP",lambda=2,mu=1,psi=3,n0=1,time=1) |>
  simulate(time=10,lambda=1) |>
  plot()

simulate("LBDP",time=4) |>
  lineages() |>
  plot()

runS2I2R2(
  time=30,
  iota2=0.01,Beta12=0.1,
  I1_0=0,I2_0=0,
  psi1=10,psi2=10,
  omega1=0.2,
  b2=0.02,d2=0.02
) |>
  freeze(seed=847110120) -> x

plot_grid(
  x |>
    plot(
      obscure=FALSE,
      palette=c("#000000ff","#440154ff","#21908cff","#ccccccff"),
      legend.position="inside",
      legend.position.inside=c(0.7,0.7)
    ),
  x |>
    lineages(obscure=FALSE) |>
    plot(
      palette=c("#000000ff","#440154ff","#21908cff","#ccccccff"),
      legend.position="none"
    ),
  ncol=1,
  align="hv",
  axis="b"
)
\dontrun{

  library(ggplot2)
  library(phylopomp)
  times <- seq(from=0,to=8,by=0.1)[-1]

  png_files <- sprintf(
    file.path(tempdir(),"frame\%05d.png"),
    seq_len(2*length(times))
  )

  pb <- utils::txtProgressBar(0,2*length(times),0,style=3)
  x <- simulate("SIIR",time=0,Beta1=5,Beta2=10,gamma=1,omega=0.5,
    psi1=0.2,psi2=0.1,sigma12=1,sigma21=1,S0=200,I1_0=3,I2_0=2)

  img <- 1
  for (k in seq.int(from=1,to=length(times),by=1)) {
    x <- simulate(x,time=times[k])
    ggsave(
      filename=png_files[img],
      plot=plot(
        x, t0=0, time=max(times),
        points=FALSE, prune=FALSE, obscure=FALSE,
        palette=c("#ffcb05","#dddddd"),
        axis.line=element_line(color="white"),
        axis.ticks=element_line(color="white"),
        axis.text=element_blank(),
        plot.background=element_rect(fill=NA,color=NA),
        panel.background=element_rect(fill=NA,color=NA)
      ),
      device="png",dpi=300,
      height=2,width=3,units="in"
    )
    setTxtProgressBar(pb,img)
    img <- img+1
  }

  for (k in seq.int(from=length(times),to=1,by=-1)) {
    x <- curtail(x,time=times[k])
    ggsave(
      filename=png_files[img],
      plot=plot(
        x, t0=0, time=max(times),
        points=FALSE, prune=FALSE, obscure=FALSE,
        palette=c("#ffcb05","#dddddd"),
        axis.line=element_line(color="white"),
        axis.ticks=element_line(color="white"),
        axis.text=element_blank(),
        plot.background=element_rect(fill=NA,color=NA),
        panel.background=element_rect(fill=NA,color=NA)
      ),
      device="png",dpi=300,
      height=2,width=3,units="in"
    )
    setTxtProgressBar(pb,img)
    img <- img+1
  }

  library(gifski)
  gif_file <- "movie1.gif"
  gifski(png_files,gif_file,delay=0.08,loop=TRUE)
  unlink(png_files)

}
}
\references{
\Yu2017
}
