% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.gpsim}
\title{Special print functions}
\usage{
\method{print}{gpsim}(x, ...)
}
\arguments{
\item{x}{\code{gpsim} object.}

\item{...}{arguments passed to print.}
}
\description{
Print functions for \sQuote{gpsim} objects.
}
\keyword{internal}
