% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si2r.R, R/si2r_pomp.R
\name{si2r}
\alias{si2r}
\alias{SI2R}
\alias{runSI2R}
\alias{continueSI2R}
\alias{si2rs_pomp}
\title{Two-deme model of superspreading}
\usage{
runSI2R(
  time,
  t0 = 0,
  Beta = 5,
  kappa = 2,
  gamma = 1,
  omega = 0,
  chi = 1,
  etaL = 1,
  etaH = 3,
  pop = 500,
  S0 = 0.98,
  IL0 = 0.02,
  IH0 = 0,
  R0 = 0
)

continueSI2R(
  object,
  time,
  Beta = NA,
  kappa = NA,
  gamma = NA,
  omega = NA,
  chi = NA,
  etaL = NA,
  etaH = NA
)

si2rs_pomp(
  x,
  Beta,
  kappa,
  gamma,
  omega,
  chi,
  etaL,
  etaH,
  S0,
  IL0,
  IH0,
  R0,
  pop
)
}
\arguments{
\item{time}{end timepoint of simulation}

\item{t0}{beginning timepoint of simulation}

\item{Beta}{transmission rate}

\item{kappa}{super-spreading transmission factor}

\item{gamma}{recovery rate}

\item{omega}{rate of waning of immunity}

\item{chi}{(destructive) sampling rate}

\item{etaL}{rate of transition from low-spreading to super-spreading behavior}

\item{etaH}{rate of transition from super-spreading to low-spreading behavior}

\item{pop}{population size}

\item{S0}{initial fraction of susceptible population}

\item{IL0}{initial fraction of low-spreading population}

\item{IH0}{initial fraction of super-spreading population}

\item{R0}{initial fraction of immune population}

\item{object}{a previously computed simulation}

\item{x}{genealogy in \pkg{phylopomp} format.}
}
\value{
\code{runSI2R} and \code{continueSI2R} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{SI2R}.

\code{si2rs_pomp} returns a \sQuote{pomp} object.
}
\description{
Deme L consists of "low-rate spreaders" that transmit at rate \code{Beta}.
Deme H consists of "superspreaders" who transmit at a higher rate
\code{kappa*Beta}.
Sampling is destructive.
}
\details{
\code{si2rs_pomp} constructs a \sQuote{pomp} object containing a given set of data and an SI2RS model.
}
\seealso{
More example genealogy processes:
\code{\link{bdei}},
\code{\link{bdss}},
\code{\link{lbdp}},
\code{\link{mers}},
\code{\link{moran}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{siir}},
\code{\link[=simulate]{simulate()}},
\code{\link{sir}},
\code{\link{strains}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
