% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siir.R
\name{siir}
\alias{siir}
\alias{SIIR}
\alias{runSIIR}
\alias{continueSIIR}
\title{Two-strain SIR model}
\usage{
runSIIR(
  time,
  t0 = 0,
  Beta1 = 5,
  Beta2 = 5,
  gamma = 1,
  psi1 = 1,
  psi2 = 0,
  sigma12 = 0,
  sigma21 = 0,
  omega = 0,
  pop = 500,
  S_0 = 0.96,
  I1_0 = 0.02,
  I2_0 = 0.02,
  R_0 = 0
)

continueSIIR(
  object,
  time,
  Beta1 = NA,
  Beta2 = NA,
  gamma = NA,
  psi1 = NA,
  psi2 = NA,
  sigma12 = NA,
  sigma21 = NA,
  omega = NA
)
}
\arguments{
\item{time}{end timepoint of simulation}

\item{t0}{beginning timepoint of simulation}

\item{Beta1}{transmission rate for strain 1}

\item{Beta2}{transmission rate for strain 2}

\item{gamma}{recovery rate}

\item{psi1}{sampling rate for deme 1}

\item{psi2}{sampling rate for deme 2}

\item{sigma12}{rate of movement from deme 1 to deme 2}

\item{sigma21}{rate of movement from deme 2 to deme 1}

\item{omega}{rate of loss of immunity}

\item{pop}{population size}

\item{S_0}{initial fraction of susceptible population}

\item{I1_0}{initial fraction of deme-1 infected population}

\item{I2_0}{initial fraction of deme-2 infected population}

\item{R_0}{initial fraction of immune population}

\item{object}{a previously computed simulation}
}
\value{
\code{runSIIR} and \code{continueSIIR} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{SIIR}.
}
\description{
Two distinct pathogen strains compete for susceptibles.
}
\seealso{
More example genealogy processes:
\code{\link{bdei}},
\code{\link{bdss}},
\code{\link{lbdp}},
\code{\link{mers}},
\code{\link{moran}},
\code{\link{mtbd2}},
\code{\link{mtbd3}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link[=simulate]{simulate()}},
\code{\link{sir}},
\code{\link{strains}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
