% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R, R/sir_pomp.R
\name{sir}
\alias{sir}
\alias{SIR}
\alias{SIRS}
\alias{runSIR}
\alias{continueSIR}
\alias{runSIRS}
\alias{continueSIRS}
\alias{sir_pomp}
\alias{sirs_pomp}
\title{Classical susceptible-infected-recovered model}
\usage{
runSIR(
  time,
  t0 = 0,
  Beta = 4,
  gamma = 1,
  psi = 1,
  chi = 0,
  omega = 0,
  pop = 100,
  S0 = 0.95,
  I0 = 0.05,
  R0 = 0
)

continueSIR(
  object,
  time,
  Beta = NA,
  gamma = NA,
  psi = NA,
  chi = NA,
  omega = NA
)

runSIRS(
  time,
  t0 = 0,
  Beta = 4,
  gamma = 1,
  psi = 1,
  chi = 0,
  omega = 0,
  pop = 100,
  S0 = 0.95,
  I0 = 0.05,
  R0 = 0
)

continueSIRS(
  object,
  time,
  Beta = NA,
  gamma = NA,
  psi = NA,
  chi = NA,
  omega = NA
)

sir_pomp(x, Beta, gamma, psi, chi = 0, omega = 0, S0, I0, R0, pop)

sirs_pomp(x, Beta, gamma, psi, chi = 0, omega = 0, S0, I0, R0, pop)
}
\arguments{
\item{time}{end timepoint of simulation}

\item{t0}{beginning timepoint of simulation}

\item{Beta}{transmission rate}

\item{gamma}{recovery rate}

\item{psi}{per capita (nondestructive) sampling rate}

\item{chi}{per capita (destructive) sampling rate}

\item{omega}{rate of waning of immunity}

\item{pop}{size of population}

\item{S0}{initial size of susceptible population}

\item{I0}{initial size of infected population}

\item{R0}{initial size of immune population}

\item{object}{a previously computed simulation}

\item{x}{genealogy in \pkg{phylopomp} format (i.e., an object that inherits from \sQuote{gpgen}).}
}
\value{
\code{runSIR} and \code{continueSIR} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{SIR}.

\code{sir_pomp} and \code{sirs_pomp} return \sQuote{pomp} objects.
}
\description{
A single, unstructured population of hosts.
}
\details{
\code{sir_pomp} constructs a \sQuote{pomp} object containing a given set of data and a SIR model.
}
\examples{
simulate("SIR",Beta=2,gamma=1,psi=2,pop=1000,I0=0.005,time=5) |>
  simulate(Beta=5,gamma=2,time=10,psi=3) |>
  plot()

runSIR(Beta=3,gamma=1,psi=2,pop=25,S0=0.8,I0=0.2,R0=0,time=5,t0=-1) |>
  plot(points=TRUE)

runSIR(Beta=3,gamma=0.1,psi=0.2,pop=100,S0=1,I0=0.05,R0=0,time=2,t0=0) -> x
plot_grid(plotlist=list(plot(x,points=TRUE),diagram(x)),
  ncol=1,rel_heights=c(4,1))

simulate("SIRS",omega=1,time=20,I0=0.05) |> plot()
simulate("SIRS",omega=1,time=20,I0=0.05) |> lineages() |> plot()
}
\references{
\King2024

\King2022
}
\seealso{
More example genealogy processes:
\code{\link{bdei}},
\code{\link{bdss}},
\code{\link{lbdp}},
\code{\link{mers}},
\code{\link{moran}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link[=simulate]{simulate()}},
\code{\link{strains}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
