% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirwr.R
\name{sirwr}
\alias{sirwr}
\alias{SIRwr}
\alias{runSIRwr}
\alias{continueSIRwr}
\title{SIR model with S, M, L segments and reassortment}
\usage{
runSIRwr(
  time,
  t0 = 0,
  Beta = 2,
  gamma = 1,
  psi = 1,
  delta = 0,
  rhoS = 0,
  rhoM = 0.1,
  rhoL = 0,
  S0 = 100,
  I0 = 2,
  R0 = 0
)

continueSIRwr(
  object,
  time,
  Beta = NA,
  gamma = NA,
  psi = NA,
  delta = NA,
  rhoS = NA,
  rhoM = NA,
  rhoL = NA
)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{Beta}{transmission rate.}

\item{gamma}{recovery rate.}

\item{psi}{sampling rate.}

\item{delta}{immunity waning rate}

\item{rhoS}{reassortment rate of S segment}

\item{rhoM}{reassortment rate of M segment}

\item{rhoL}{reassortment rate of L segment}

\item{S0}{initial size of susceptible population.}

\item{I0}{initial size of infected population.}

\item{R0}{initial size of recovered population.}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}
}
\value{
An object of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{SIR}.
}
\description{
A single, unstructured population of hosts.
}
\examples{
simulate("SIRwr",Beta=2,gamma=1,psi=2,
  rhoS=0, rhoM=0, rhoL=0,
  S0=1000,I0=5,time=5) |>
  simulate(Beta=5,gamma=2,time=10,psi=3) |>
  plot()

runSIRwr(Beta=3,gamma=1,psi=2,S0=20,I0=5,R0=0,time=5,t0=-1) |>
  plot(points=TRUE)

runSIRwr(Beta=3,gamma=0.1,psi=0.2,S0=100,I0=5,R0=0,time=2,t0=0) -> x
plot_grid(plotlist=list(plot(x,points=TRUE),diagram(x)),
  ncol=1,rel_heights=c(4,1))

simulate("SIRwr",delta=1,time=20,I0=4) |> plot()
simulate("SIRwr",delta=1,time=20,I0=4) |> lineages() |> plot()
}
\seealso{
More example genealogy processes: 
\code{\link{lbdp}},
\code{\link{moran}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sir}}
}
\concept{Genealogy processes}
