% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirws.R
\name{sirws}
\alias{sirws}
\alias{playSIRwS}
\alias{SIRwS}
\alias{getInfo.SIRwS_gpsim}
\title{SIR with sampling simulator.}
\usage{
playSIRwS(
  data = NULL,
  ...,
  beta,
  gamma,
  psi,
  iota,
  S0,
  I0,
  t0 = 0,
  times,
  tree = FALSE
)

\method{getInfo}{SIRwS_gpsim}(data, prune = TRUE, tree = TRUE)
}
\arguments{
\item{data}{optional data frame; output from \code{playSIRwS}.}

\item{...}{additional arguments; ignored.}

\item{beta}{transmission rate}

\item{gamma}{recovery rate}

\item{psi}{sampling rate}

\item{iota}{import rate}

\item{S0}{initial size of susceptible population}

\item{I0}{initial size of infected population}

\item{t0}{initial time}

\item{times}{times at which output is requested.}

\item{tree}{logical; represent the genealogical tree in Newick format?}

\item{prune}{Prune the tree?}
}
\value{
A \code{tibble} with \code{state} attribute.
}
\description{
Run the simulator.
}
\examples{
tout <- seq(0,5,by=1)
playSIRwS(beta=2,gamma=1,psi=2,iota=0,S0=1000,I0=5,times=tout,t0=0,tree=TRUE) -> x

playSIRwS(x,times=6:10,psi=0.1,tree=TRUE) -> x
treeplot(x)

y <- getInfo(x)
treeplot(y,points=TRUE)

y[c("etimes","lineages")] \%>\%
  as_tibble() \%>\%
  ggplot(aes(x=etimes,y=lineages))+
  geom_step()+
  labs(x="time")
}
\seealso{
Other Genealogy processes: 
\code{\link{moran}}
}
\concept{Genealogy processes}
