% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\title{Subsample a genealogy}
\usage{
subsample(object, frac)
}
\arguments{
\item{object}{a \sQuote{gpgen} object, possibly with \sQuote{model} attribute.}

\item{frac}{fraction of samples to retain}
}
\value{
A bare genealogy object containing the randomly subsampled genealogy.
}
\description{
Drops a random selection of samples
}
