% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplot.R
\name{treeplot}
\alias{treeplot}
\title{Fancy tree plotter}
\usage{
treeplot(
  tree,
  time = NULL,
  illus = NULL,
  root_time = 0,
  ladderize = TRUE,
  points = FALSE,
  diagram = FALSE
)
}
\arguments{
\item{tree}{character; tree representation in Newick format.}

\item{time}{numeric; times of the genealogies.}

\item{illus}{character; genealogy process diagram information.}

\item{root_time}{numeric; time of the root.}

\item{ladderize}{Ladderize?}

\item{points}{Show nodes and tips?}

\item{diagram}{Show a diagram?}
}
\value{
A printable \code{ggtree} object.
}
\description{
Plots a genealogical tree.
}
