% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{yaml}
\alias{yaml}
\title{YAML output}
\usage{
yaml(object)
}
\arguments{
\item{object}{\code{gpsim} object.}
}
\value{
A string in yaml format, with class \dQuote{gpyaml}.
}
\description{
Human- and machine-readable description
}
\examples{
simulate("SIIR",time=1) |> yaml()

}
