% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdei.R
\name{bdei}
\alias{bdei}
\alias{BDEI}
\alias{runBDEI}
\alias{continueBDEI}
\title{Linear birth-death with exposed and infectious classes}
\usage{
runBDEI(
  time,
  t0 = 0,
  mu = 1,
  la = 0.5,
  psi = 0.25,
  p = 0.3,
  IE0 = 0,
  II0 = 1
)

continueBDEI(
  object,
  time,
  mu = NA,
  la = NA,
  psi = NA,
  p = NA
)
}
\arguments{
\item{time}{end timepoint of simulation}

\item{t0}{beginning timepoint of simulation}

\item{mu}{per capita progression rate from exposed to infectious}

\item{la}{per capita transmission rate from infectious to exposed}

\item{psi}{per capita removal rate of infectious lineages}

\item{p}{probability of sampling upon infectious removal}

\item{IE0}{initial number of exposed lineages}

\item{II0}{initial number of infectious lineages}

\item{object}{a previously computed simulation}
}
\value{
\code{runBDEI} and \code{continueBDEI} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{BDEI}.
}
\description{
Two-deme linear birth-death process with exposed (incubating) and infectious lineages.
Transmission creates exposed offspring from infectious donors; exposed lineages progress to infectious.
Removal and Stadler sampling-on-removal apply to infectious lineages only.
Compatible with the BDEI model of Voznica et al. and structured linear birth-death theory.
}
\seealso{
More example genealogy processes:
\code{\link{bdss}},
\code{\link{lbdp}},
\code{\link{mers}},
\code{\link{moran}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sir}},
\code{\link{strains}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
