% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdei.R
\name{bdei}
\alias{bdei}
\alias{BDEI}
\alias{runBDEI}
\alias{continueBDEI}
\title{Linear birth-death with exposed and infectious classes}
\usage{
runBDEI(
  time,
  t0 = 0,
  mu = 0.128205128205128,
  lambda_ie = 0.766666666666667,
  psi = 0.333333333333333,
  p = 0.505,
  pop = 1,
  fe = 0,
  fi = 1
)

continueBDEI(object, time, mu = NA, lambda_ie = NA, psi = NA, p = NA)
}
\arguments{
\item{time}{end timepoint of simulation}

\item{t0}{beginning timepoint of simulation}

\item{mu}{per capita progression rate from exposed to infectious}

\item{lambda_ie}{per capita transmission rate from infectious to exposed}

\item{psi}{per capita removal rate of infectious lineages}

\item{p}{probability of sampling upon infectious removal}

\item{pop}{reference population size for scaling initial lineage counts}

\item{fe}{initial fraction of exposed lineages (ie/pop)}

\item{fi}{initial fraction of infectious lineages (ii/pop)}

\item{object}{a previously computed simulation}
}
\value{
\code{runBDEI} and \code{continueBDEI} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{BDEI}.
}
\description{
Two-deme linear birth-death process with exposed (incubating) and infectious lineages.
Transmission creates exposed offspring from infectious donors; exposed lineages progress to infectious.
Removal and Stadler sampling-on-removal apply to infectious lineages only.
Compatible with the BDEI model of Voznica et al. (2022), https://doi.org/10.1038/s41467-022-31511-0,
and structured linear birth-death theory (Stadler 2010, https://doi.org/10.1016/j.jtbi.2010.09.010;
King et al. 2024, https://arxiv.org/abs/2405.17032).
YAML rate defaults are midpoints of the PhyloDeep training priors
(R0 in [1.1, 3.5], T_I in [1, 5], T_E/T_I in [0.2, 5], p in [0.01, 1]).
}
\seealso{
More example genealogy processes:
\code{\link{bdss}},
\code{\link{lbdp}},
\code{\link{mers}},
\code{\link{moran}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sir}},
\code{\link{strains}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
