% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdss.R, R/bdss_pomp.R
\name{bdss}
\alias{bdss}
\alias{BDSS}
\alias{runBDSS}
\alias{continueBDSS}
\alias{bdss_pomp}
\title{Linear birth-death with superspreading}
\usage{
runBDSS(
  time,
  t0 = 0,
  lambda_nn = 0.875,
  lambda_ns = 0.125,
  lambda_sn = 5.5,
  lambda_ss = 0.75,
  mu = 0.5,
  chi = 0.5,
  pop = 1,
  N0 = 1,
  S0 = 0
)

continueBDSS(
  object,
  time,
  lambda_nn = NA,
  lambda_ns = NA,
  lambda_sn = NA,
  lambda_ss = NA,
  mu = NA,
  chi = NA
)

bdss_pomp(x, lambda_nn, lambda_ns, lambda_sn, lambda_ss, mu, chi, N0, S0, pop)
}
\arguments{
\item{time}{end timepoint of simulation}

\item{t0}{beginning timepoint of simulation}

\item{lambda_nn}{normal-to-normal transmission rate}

\item{lambda_ns}{normal-to-superspreader transmission rate}

\item{lambda_sn}{superspreader-to-normal transmission rate}

\item{lambda_ss}{superspreader-to-superspreader transmission rate}

\item{mu}{per capita removal (death or recovery) rate}

\item{chi}{per capita destructive sampling rate}

\item{pop}{initial population size}

\item{N0}{initial fraction of normal spreaders}

\item{S0}{initial fraction of superspreaders}

\item{object}{a previously computed simulation}

\item{x}{genealogy in \pkg{phylopomp} format.}
}
\value{
\code{runBDSS} and \code{continueBDSS} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{BDSS}.

\code{bdss_pomp} returns a \sQuote{pomp} object.
}
\description{
Two-deme linear birth-death process with heterogeneous per-lineage transmission rates.
Identical to the BDSS model of Voznica et al. (2022, https://doi.org/10.1038/s41467-022-31511-0), though the parameterization differs.
}
\details{
\code{bdss_pomp} constructs a \sQuote{pomp} object containing a given set of data and a BDSS model.
}
\seealso{
More example genealogy processes:
\code{\link{bdei}},
\code{\link{lbdp}},
\code{\link{mers}},
\code{\link{moran}},
\code{\link{mtbd2}},
\code{\link{mtbd3}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link[=simulate]{simulate()}},
\code{\link{sir}},
\code{\link{strains}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
