% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdss.R
\name{bdss}
\alias{bdss}
\alias{BDSS}
\alias{runBDSS}
\alias{continueBDSS}
\title{Linear birth-death with superspreading}
\usage{
runBDSS(
  time,
  t0 = 0,
  la_nn = 0.1,
  la_ns = 0.3,
  la_sn = 0.5,
  la_ss = 1.5,
  mu = 0.25,
  p = 0.3,
  IN0 = 1,
  IS0 = 0
)

continueBDSS(
  object,
  time,
  la_nn = NA,
  la_ns = NA,
  la_sn = NA,
  la_ss = NA,
  mu = NA,
  p = NA
)
}
\arguments{
\item{time}{end timepoint of simulation}

\item{t0}{beginning timepoint of simulation}

\item{la_nn}{normal-to-normal transmission rate}

\item{la_ns}{normal-to-superspreader transmission rate}

\item{la_sn}{superspreader-to-normal transmission rate}

\item{la_ss}{superspreader-to-superspreader transmission rate}

\item{mu}{per capita removal rate}

\item{p}{probability of sampling upon removal}

\item{IN0}{initial number of normal spreaders}

\item{IS0}{initial number of superspreaders}

\item{object}{a previously computed simulation}
}
\value{
\code{runBDSS} and \code{continueBDSS} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{BDSS}.
}
\description{
Two-deme linear birth-death process with heterogeneous per-lineage transmission rates.
Compatible with the BDSS model of Voznica et al. and structured birth-death models of Stadler.
}
\seealso{
More example genealogy processes:
\code{\link{lbdp}},
\code{\link{mers}},
\code{\link{moran}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sir}},
\code{\link{strains}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
