% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdss.R
\name{bdss}
\alias{bdss}
\alias{BDSS}
\alias{runBDSS}
\alias{continueBDSS}
\title{Linear birth-death with superspreading}
\usage{
runBDSS(
  time,
  t0 = 0,
  lambda_nn = 0.875,
  lambda_ns = 0.125,
  lambda_sn = 5.6875,
  lambda_ss = 0.8125,
  mu = 1,
  p = 0.505,
  pop = 1,
  fn = 1,
  fs = 0
)

continueBDSS(
  object,
  time,
  lambda_nn = NA,
  lambda_ns = NA,
  lambda_sn = NA,
  lambda_ss = NA,
  mu = NA,
  p = NA
)
}
\arguments{
\item{time}{end timepoint of simulation}

\item{t0}{beginning timepoint of simulation}

\item{lambda_nn}{normal-to-normal transmission rate}

\item{lambda_ns}{normal-to-superspreader transmission rate}

\item{lambda_sn}{superspreader-to-normal transmission rate}

\item{lambda_ss}{superspreader-to-superspreader transmission rate}

\item{mu}{per capita removal rate}

\item{p}{probability of sampling upon removal}

\item{pop}{reference population size for scaling initial lineage counts}

\item{fn}{initial fraction of normal spreaders (ino/pop)}

\item{fs}{initial fraction of superspreaders (iso/pop)}

\item{object}{a previously computed simulation}
}
\value{
\code{runBDSS} and \code{continueBDSS} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{BDSS}.
}
\description{
Two-deme linear birth-death process with heterogeneous per-lineage transmission rates.
Compatible with the BDSS model of Voznica et al. (2022), https://doi.org/10.1038/s41467-022-31511-0,
and structured linear birth-death theory (Stadler 2010, https://doi.org/10.1016/j.jtbi.2010.09.010;
King et al. 2024, https://arxiv.org/abs/2405.17032).
YAML rate defaults are midpoints of the PhyloDeep training priors
(f in [0.05, 0.20], X in [3, 10], lambda_n in [0.5, 1.5], mu in [0.5, 1.5], p in [0.01, 1]).
}
\seealso{
More example genealogy processes:
\code{\link{bdei}},
\code{\link{lbdp}},
\code{\link{mers}},
\code{\link{moran}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sir}},
\code{\link{strains}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
