% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cblv.R
\name{cblv}
\alias{cblv}
\title{Compact, bijective, ladderized vector representation of a genealogy}
\usage{
cblv(object)
}
\arguments{
\item{object}{a \sQuote{gpgen} object, possibly with \sQuote{model} attribute.}
}
\value{
A matrix with two columns.
}
\description{
Computes the CBLV representation.
}
