% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curtail.R
\name{curtail}
\alias{curtail}
\title{Curtail a genealogy to the given time}
\usage{
curtail(object, time = NA, prune = TRUE, obscure = TRUE)
}
\arguments{
\item{object}{\code{gpsim} object.}

\item{time}{logical; return the current time?}

\item{prune}{logical; prune the genealogy?}

\item{obscure}{logical; obscure the demes?}
}
\value{
A curtailed genealogy object.
}
\description{
Discards all nodes beyond the given time.
}
\examples{
simulate("SIIR",time=1) |> newick()
}
