% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{diagram}
\alias{diagram}
\title{Genealogy process diagram}
\usage{
diagram(illustration, ..., gp = gpar())
}
\arguments{
\item{illustration}{character;
illustrations produced by \code{\link{getInfo}} or one of the \code{playX} functions.}

\item{...}{additional arguments; ignored.}

\item{gp}{An object of class \code{gpar}, typically the output from a call to the function \code{\link[grid:gpar]{gpar}}.
This is basically a list of graphical parameter settings.}
}
\value{
A list of \pkg{grid} graphics objects (\code{grob}s), invisibly.
}
\description{
Produces a diagram of the genealogy process state.
}
