% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{diagram}
\alias{diagram}
\title{Genealogy process diagram}
\usage{
diagram(data, prune = TRUE, m = NULL, n = NULL, ...)
}
\arguments{
\item{data}{\code{gpsim} object.}

\item{prune}{logical; prune the genealogy?}

\item{m}{width of plotting window, in nodes.
By default, the nodes will be adjusted in width to fit the window.}

\item{n}{height of the pockets, in balls.
By default, the balls will be adjusted in size to fit the space available.}

\item{...}{graphical parameter settings, suitable for passing to \code{\link[grid:gpar]{gpar}}.}
}
\value{
A \pkg{grid} graphics object (\code{grob}), invisibly.
}
\description{
Produces a diagram of the genealogy process state.
}
\examples{
library(cowplot)
runSIR(Beta=3,gamma=0.1,psi=0.2,S0=100,I0=5,R0=0,time=2,t0=0) -> x
plot(x,points=TRUE,prune=FALSE,compact=FALSE)
plot_grid(plotlist=list(plot(x,points=TRUE)[[1]],diagram(x)),
  ncol=1,rel_heights=c(4,1))
}
