% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{diagram}
\alias{diagram}
\alias{print.gpdiag}
\title{Genealogy process diagram}
\usage{
diagram(
  object,
  prune = TRUE,
  obscure = TRUE,
  m = NULL,
  n = NULL,
  ...,
  digits = 1
)

\method{print}{gpdiag}(x, newpage = is.null(vp), vp = NULL, ...)
}
\arguments{
\item{object}{\code{gpsim} object.}

\item{prune}{logical; prune the genealogy?}

\item{obscure}{logical; obscure the demes?}

\item{m}{width of plotting window, in nodes.
By default, the nodes will be adjusted in width to fit the window.}

\item{n}{height of the pockets, in balls.
By default, the balls will be adjusted in size to fit the space available.}

\item{...}{other arguments, ignored.}

\item{digits}{non-negative integer;
number of decimal digits to print in the node time}

\item{x}{ An \R object. }

\item{newpage}{draw new empty page first?}

\item{vp}{viewport to draw plot in}
}
\value{
A \pkg{grid} graphics object (\code{grob}), invisibly.
}
\description{
Produces a diagram of the genealogy process state.
}
\examples{
runSIR(Beta=3,gamma=0.1,psi=0.2,S0=100,I0=5,R0=0,time=2,t0=0) -> x
plot(x,points=TRUE,prune=FALSE)
plot_grid(plotlist=list(plot(x,points=TRUE)[[1]],diagram(x)),
  ncol=1,rel_heights=c(4,1))
}
