% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{diagram}
\alias{diagram}
\title{Genealogy process diagram}
\usage{
diagram(illustration, ...)
}
\arguments{
\item{illustration}{character;
illustrations produced by \code{\link{getInfo}} or one of the \code{playX} functions.}

\item{...}{graphical parameter settings. See \code{\link[grid:gpar]{gpar}}.}
}
\value{
A list of \pkg{grid} graphics objects (\code{grob}s), invisibly.
}
\description{
Produces a diagram of the genealogy process state.
}
