% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendat.R
\name{gendat}
\alias{gendat}
\title{Genealogy as a data frame}
\usage{
gendat(object)
}
\arguments{
\item{object}{a \sQuote{gpgen} object.}
}
\value{
A tibble.
}
\description{
Converts a given genealogy to a data frame.
}
