% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{geneal_sum}
\alias{geneal_sum}
\alias{+.gpgen}
\title{Combine genealogies}
\usage{
\method{+}{gpgen}(x, y)

geneal_sum(...)
}
\arguments{
\item{x, y, ...}{genealogies or lists of genealogies to be summed.}
}
\value{
combined genealogy.
}
\description{
A summing operator
}
