% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getinfo.R
\name{getInfo}
\alias{getInfo}
\alias{getInfo.gpsim}
\title{Retrieve information from genealogy process simulation}
\usage{
getInfo(data, ...)

\method{getInfo}{gpsim}(data, ..., prune = TRUE, compact = TRUE)
}
\arguments{
\item{data}{\code{gpsim} object.}

\item{...}{arguments passed to specific methods.}

\item{prune}{logical; prune the tree?}

\item{compact}{logical; return the tree in compact representation?}
}
\value{
A list containing the following elements
\describe{
  \item{t0}{the initial time}
  \item{time}{the current time}
  \item{tree}{the genealogical tree, in Newick format}
  \item{description}{a human readable description of the state of the genealogy process}
  \item{yaml}{the state of the genealogy process in YAML format}
  \item{structure}{the state of the genealogy process in \R list format}
  \item{lineages}{a \code{\link[tibble]{tibble}} containing the lineage count function through time}
}
}
\description{
getInfo
}
\examples{
playSIIR(Beta1=2,Beta2=4,gamma=1,psi=5,S0=200,I1_0=5,I2_0=5,R0=0,times=0:3,t0=0) |>
  continue(Beta1=5,Beta2=10,gamma=2,times=4:10,psi=10) |>
  getInfo() |>
  plot()

playSIIR(Beta1=10,Beta2=8,gamma=1,psi=2,S0=300,I1_0=2,I2_0=20,R0=0,times=0,t0=-1) |>
  continue(psi=10,times=1:5) |>
  getInfo(prune=FALSE) -> y
plot(y,points=TRUE)

library(ggplot2)
y$lineages |>
  ggplot(aes(x=time,y=lineages))+
  geom_step()
}
