% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getinfo.R
\name{getInfo}
\alias{getInfo}
\alias{getinfo}
\title{Retrieve information from genealogy process simulation}
\usage{
getInfo(data, ...)
}
\description{
getInfo
}
\examples{
playMoran(n=5,mu=choose(5,2),times=1:10,t0=0,tree=TRUE) -> x
playMoran(x,times=11:20,tree=TRUE) -> x
treeplot(x,points=TRUE)

y <- getInfo(x,prune=FALSE)
treeplot(y,points=TRUE)
}
