% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getinfo.R
\name{getInfo}
\alias{getInfo}
\alias{getinfo}
\alias{getInfo.gpsim}
\title{Retrieve information from genealogy process simulation}
\usage{
getInfo(data, ...)

\method{getInfo}{gpsim}(data, ..., prune = TRUE)
}
\arguments{
\item{data}{\code{gpsim} object.}

\item{...}{arguments passed to specific methods.}

\item{prune}{logical; prune the tree?}
}
\description{
getInfo
}
\examples{
playMoran(n=5,mu=5,times=0:10,t0=0,tree=TRUE,ill=TRUE) -> x
playMoran(x,times=11:20,tree=TRUE) -> x
plot(x)

playMoran(n=20,mu=20,times=0:20,stationary=FALSE,tree=TRUE,ill=TRUE) -> x
plot(x)

y <- getInfo(x,prune=FALSE)
plot(y,points=TRUE)
}
