% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getinfo.R
\name{getInfo}
\alias{getInfo}
\alias{getinfo}
\alias{getInfo.gpsim}
\title{Retrieve information from genealogy process simulation}
\usage{
getInfo(data, ...)

\method{getInfo}{gpsim}(data, ..., prune = TRUE)
}
\arguments{
\item{data}{\code{gpsim} object.}

\item{...}{arguments passed to specific methods.}

\item{prune}{logical; prune the tree?}
}
\description{
getInfo
}
\examples{
playMoran(n=5,mu=5,times=0:10,t0=0,tree=TRUE,ill=TRUE) -> x
playMoran(x,times=11:20,tree=TRUE) -> x
plot(x)

playMoran(n=20,mu=20,times=0:20,stationary=FALSE,tree=TRUE,ill=TRUE) -> x
plot(x)

y <- getInfo(x,prune=FALSE)
plot(y,points=TRUE)

library(tidyverse)
library(cowplot)
playMoran(n=5,mu=10,times=c(0,seq(100,200,by=25)),
  stationary=FALSE,ill=TRUE,tree=TRUE) \%>\%
  mutate(grob=diagram(illustration,fontsize=9)) -> x
plot(x[5,],points=TRUE)[[1]]+
  annotation_custom(x$grob[[5]],xmin=0,xmax=150,ymin=12,ymax=16)

playMoran(n=5,mu=5,times=0:3,stationary=TRUE,tree=TRUE,ill=TRUE) \%>\%
  mutate(grob=diagram(illustration)) -> x
plot(x,points=TRUE)
cowplot::plot_grid(plotlist=x$grob,ncol=1)

playMoran(n=5,mu=5,times=0:3,stationary=FALSE,tree=TRUE,ill=TRUE) \%>\%
  mutate(grob=diagram(illustration)) -> x
plot(x,points=TRUE)
}
