% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getinfo.R
\name{getInfo}
\alias{getInfo}
\title{getInfo}
\usage{
getInfo(
  object,
  prune = TRUE,
  t0 = FALSE,
  time = FALSE,
  description = FALSE,
  structure = FALSE,
  yaml = FALSE,
  lineages = FALSE,
  tree = FALSE,
  compact = TRUE
)
}
\arguments{
\item{object}{\code{gpsim} object.}

\item{prune}{logical; prune the genealogy?}

\item{t0}{logical; return the zero-time?}

\item{time}{logical; return the current time?}

\item{description}{logical; return the description?}

\item{structure}{logical; return the structure in \R list format?}

\item{yaml}{logical; return the structure in YAML format?}

\item{lineages}{logical; return the lineage-count function?}

\item{tree}{logical; return the tree?}

\item{compact}{logical; return the tree in compact representation?}
}
\value{
A list containing the requested elements, including any or all of:
\describe{
  \item{t0}{the initial time}
  \item{time}{the current time}
  \item{tree}{the genealogical tree, in Newick format}
  \item{description}{a human readable description of the state of the genealogy process}
  \item{yaml}{the state of the genealogy process in YAML format}
  \item{structure}{the state of the genealogy process in \R list format}
  \item{lineages}{a \code{\link[tibble]{tibble}} containing the lineage count function through time}
}
}
\description{
Retrieve information from genealogy process simulation
}
\examples{
simulate("SIIR",Beta1=2,Beta2=4,gamma=1,psi=5,S0=200,I1_0=5,I2_0=5,time=3,t0=0) |>
  continue(Beta1=5,Beta2=10,gamma=2,time=10,psi=10) |>
  plot()

runSIIR(Beta1=10,Beta2=8,gamma=1,psi=2,S0=300,I1_0=2,I2_0=20,R0=0,time=0,t0=-1) |>
  continue(psi=10,time=2) |>
  plot(points=TRUE,prune=FALSE)

library(ggplot2)
simulate("SIIR",Beta1=10,Beta2=8,gamma=1,psi=2,S0=300,I1_0=2,I2_0=20,time=5) |>
  lineages() |>
  ggplot(aes(x=time,y=lineages))+
  geom_step()
}
