% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R, R/print.R
\name{print.gpdiag}
\alias{print.gpdiag}
\alias{genealogyGrob}
\alias{nodeGrob}
\alias{pocketGrob}
\alias{ballGrob}
\alias{resizingTextGrob}
\alias{drawDetails.resizingTextGrob}
\alias{preDrawDetails.resizingTextGrob}
\alias{postDrawDetails.resizingTextGrob}
\alias{drawDetails.ballGrob}
\alias{preDrawDetails.ballGrob}
\alias{postDrawDetails.ballGrob}
\alias{print.gpsim}
\title{Diagramming internals}
\usage{
\method{print}{gpdiag}(x, newpage = is.null(vp), vp = NULL, ...)

genealogyGrob(data, m = NULL, n = NULL, vp = NULL)

nodeGrob(data, digits = 1, n = NULL, vp = NULL)

pocketGrob(data, n = NULL, vp = NULL)

ballGrob(data, vp = NULL)

resizingTextGrob(..., vp = NULL)

\method{drawDetails}{resizingTextGrob}(x, recording = TRUE)

\method{preDrawDetails}{resizingTextGrob}(x)

\method{postDrawDetails}{resizingTextGrob}(x)

\method{drawDetails}{ballGrob}(x, recording = TRUE)

\method{preDrawDetails}{ballGrob}(x)

\method{postDrawDetails}{ballGrob}(x)

\method{print}{gpsim}(x, ...)
}
\arguments{
\item{x}{ An \R object. }

\item{vp}{a \code{\link[grid]{viewport}} object (or \code{NULL}).}

\item{...}{arguments to be passed to \code{\link[grid:textGrob]{textGrob}}.}

\item{data}{list; pocket structure}

\item{m}{width of plotting window, in nodes.
By default, the nodes will be adjusted in width to fit the window.}

\item{n}{length of longest genealogy}

\item{digits}{non-negative integer;
number of decimal digits to print in the node time}

\item{recording}{ A logical value indicating whether a grob
  is being added to the display list or redrawn from the display list. }
}
\description{
Facilities to produce diagrammatic representations
of genealogy process states.
}
\details{
Code for the resizing text adapted from a blog post by Mark Heckmann
(https://ryouready.wordpress.com/2012/08/01/creating-a-text-grob-that-automatically-adjusts-to-viewport-size/).
}
\keyword{internal}
\keyword{internals}
