% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R
\name{tableauGrob}
\alias{tableauGrob}
\alias{playerGrob}
\alias{resizingTextGrob}
\alias{drawDetails.resizingTextGrob}
\alias{preDrawDetails.resizingTextGrob}
\alias{postDrawDetails.resizingTextGrob}
\alias{ballGrob}
\alias{drawDetails.ballGrob}
\alias{preDrawDetails.ballGrob}
\alias{postDrawDetails.ballGrob}
\title{Diagramming internals}
\usage{
tableauGrob(data, n, vp = NULL)

playerGrob(name, ballA, ballAcol, ballB, ballBcol, slate, vp = NULL)

resizingTextGrob(y, ..., vp = NULL)

\method{drawDetails}{resizingTextGrob}(x, recording = TRUE)

\method{preDrawDetails}{resizingTextGrob}(x)

\method{postDrawDetails}{resizingTextGrob}(x)

ballGrob(y, label, color, ..., vp = NULL)

\method{drawDetails}{ballGrob}(x, recording = TRUE)

\method{preDrawDetails}{ballGrob}(x)

\method{postDrawDetails}{ballGrob}(x)
}
\arguments{
\item{data}{illustration vector}

\item{n}{length of longest tableau}

\item{vp}{a \code{\link[grid]{viewport}} object (or \code{NULL}).}

\item{name}{player name}

\item{ballA, ballB}{ball names}

\item{ballAcol, ballBcol}{ball colors}

\item{slate}{time}

\item{y}{height}

\item{...}{arguments to be passed to \code{\link[grid:textGrob]{textGrob}}.}

\item{x}{ An \R object. }

\item{recording}{ A logical value indicating whether a grob
  is being added to the display list or redrawn from the display list. }

\item{label}{ball name}

\item{color}{ball color}
}
\description{
Facilities to produce diagrammatic representations
of genealogy process states.
}
\details{
Code for the resizing text adapted from a blog post by Mark Heckmann
(https://ryouready.wordpress.com/2012/08/01/creating-a-text-grob-that-automatically-adjusts-to-viewport-size/).
}
\keyword{internal}
