% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R, R/print.R
\name{genealogy diagram internals}
\alias{genealogy diagram internals}
\alias{genealogyGrob}
\alias{nodeGrob}
\alias{pocketGrob}
\alias{ballGrob}
\alias{resizingTextGrob}
\alias{drawDetails.resizingTextGrob}
\alias{preDrawDetails.resizingTextGrob}
\alias{postDrawDetails.resizingTextGrob}
\alias{drawDetails.ballGrob}
\alias{preDrawDetails.ballGrob}
\alias{postDrawDetails.ballGrob}
\alias{print.gpsim}
\alias{print.gpyaml}
\title{Diagramming internals}
\usage{
genealogyGrob(object, m = NULL, n = NULL, vp = NULL, palette, ...)

nodeGrob(object, digits = 1, palette, n = NULL, vp = NULL)

pocketGrob(object, n, vp = NULL)

ballGrob(object, vp = NULL)

resizingTextGrob(..., vp = NULL)

\method{drawDetails}{resizingTextGrob}(x, recording = TRUE)

\method{preDrawDetails}{resizingTextGrob}(x)

\method{postDrawDetails}{resizingTextGrob}(x)

\method{drawDetails}{ballGrob}(x, recording = TRUE)

\method{preDrawDetails}{ballGrob}(x)

\method{postDrawDetails}{ballGrob}(x)

\method{print}{gpsim}(x, ...)

\method{print}{gpyaml}(x, ...)
}
\arguments{
\item{object}{list; pocket structure}

\item{m}{width of plotting window, in nodes.
By default, the nodes will be adjusted in width to fit the window.}

\item{n}{length of longest genealogy}

\item{vp}{viewport to draw plot in}

\item{palette}{color palette for indicating demes.
This can be furnished either as a function or a vector of colors.
If this is a function, it should take a single integer argument, the number of colors required.
If it is a vector, it should have at least as many elements as there are demes in the genealogy.}

\item{...}{arguments to be passed to \code{\link[grid:textGrob]{textGrob}}.}

\item{digits}{non-negative integer;
number of decimal digits to print in the node time}

\item{x}{ An \R object. }

\item{recording}{ A logical value indicating whether a grob
  is being added to the display list or redrawn from the display list. }
}
\description{
Facilities to produce diagrammatic representations
of genealogy process states.
}
\details{
Code for the resizing text adapted from a blog post by Mark Heckmann
(https://ryouready.wordpress.com/2012/08/01/creating-a-text-grob-that-automatically-adjusts-to-viewport-size/).
}
\keyword{internal}
\keyword{internals}
