% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbdp.R, R/lbdp_pomp.R, R/lbdp_exact.R
\name{lbdp}
\alias{lbdp}
\alias{LBDP}
\alias{runLBDP}
\alias{continueLBDP}
\alias{lbdp_pomp}
\alias{lbdp_exact}
\title{Linear birth-death-sampling model}
\usage{
runLBDP(
  time,
  t0 = 0,
  lambda = 2,
  mu = 1,
  psi = 1,
  chi = 0,
  n0 = 5,
  max_sample = 10000
)

continueLBDP(object, time, lambda = NA, mu = NA, psi = NA, chi = NA)

lbdp_pomp(x, lambda, mu, psi, chi = 0, n0 = 1)

lbdp_exact(x, lambda, mu, psi, chi = 0, n0 = 1)
}
\arguments{
\item{time}{end timepoint of simulation}

\item{t0}{beginning timepoint of simulation}

\item{lambda}{per capita birth rate}

\item{mu}{per capita death rate}

\item{psi}{per capita non-destructive sampling rate}

\item{chi}{per capita destructive sampling rate}

\item{n0}{population size at time t0}

\item{max_sample}{maximum number of samples}

\item{object}{a previously computed simulation}

\item{x}{genealogy in \pkg{phylopomp} format (i.e., an object that inherits from \sQuote{gpgen}).}
}
\value{
\code{runLBDP} and \code{continueLBDP} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{LBDP}.

\code{lbdp_exact} returns the log likelihood of the genealogy.
Note that the time since the most recent sample is informative.
}
\description{
The genealogy process induced by a simple linear birth-death process with constant-rate sampling.
}
\details{
\code{lbdp_pomp} constructs a \pkg{pomp} object containing a given set of data and a linear birth-death-sampling process.

\code{lbdp_exact} gives the exact log likelihood of a linear birth-death process with (optionally destructive) sampling, conditioned on the population size at time 0.
If \code{n0=NA}, then the log likelihood is maximized over the initial population size.
}
\examples{
simulate("LBDP",time=4) |> plot(points=TRUE)

simulate("LBDP",lambda=2,mu=1,psi=3,n0=1,time=1) |>
  simulate(time=10,lambda=1) |>
  plot()

simulate("LBDP",time=4) |>
  lineages() |>
  plot()

}
\references{
\King2024

\King2022

\Stadler2010

\Stadler2010

\King2024
}
\seealso{
More example genealogy processes:
\code{\link{bdei}},
\code{\link{bdss}},
\code{\link{mers}},
\code{\link{moran}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link[=simulate]{simulate()}},
\code{\link{sir}},
\code{\link{strains}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
