% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbdp.R
\name{lbdp}
\alias{lbdp}
\alias{LBDP}
\alias{runLBDP}
\alias{continueLBDP}
\title{Linear birth-death-sampling model}
\usage{
runLBDP(time, t0 = 0, lambda = 2, mu = 1, psi = 1, n0 = 5)

continueLBDP(object, time, lambda = NA, mu = NA, psi = NA)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{lambda}{per capita birth rate}

\item{mu}{per capita recovery rate.}

\item{psi}{per capita sampling rate.}

\item{n0}{initial population size}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}
}
\value{
An object of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{LBDP}.
}
\description{
The genealogy process induced by a simple linear birth-death process
with constant-rate sampling.
}
\examples{
simulate("LBDP",time=4) |> plot(points=TRUE)

simulate("LBDP",lambda=2,mu=1,psi=3,n0=1,time=1) |>
  simulate(time=10,lambda=1) |>
  plot()

simulate("LBDP",time=4) |>
  lineages() |>
  plot()

}
\seealso{
More example genealogy processes: 
\code{\link{moran}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sir}}
}
\concept{Genealogy processes}
