% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbdp.R
\name{lbdp}
\alias{lbdp}
\alias{LBDP}
\alias{playLBDP}
\title{Linear birth-death process.}
\usage{
playLBDP(data = NULL, ..., lambda, mu, n0, psi, t0 = 0, times, tree = FALSE)
}
\arguments{
\item{data}{optional data frame;
output from \code{playLBDP}.}

\item{...}{additional arguments; ignored.}

\item{lambda}{birth rate}

\item{mu}{death rate}

\item{n0}{initial population size}

\item{psi}{sampling rate}

\item{t0}{initial time}

\item{times}{times at which output is requested.}

\item{tree}{logical; represent the genealogical tree in Newick format?}
}
\value{
A \code{tibble} with \code{state} attribute.
}
\description{
Run the LBDP simulator.
}
\examples{
playLBDP(lambda=1,mu=0.5,psi=1,n0=3,times=seq(0,4,by=0.2),tree=TRUE) -> x
plot(x,points=TRUE)

y <- getInfo(x,prune=FALSE)
plot(y,points=TRUE)

library(ggplot2)
y$lineages \%>\%
  ggplot(aes(x=time,y=lineages))+
  geom_step()+
  geom_vline(xintercept=y$etimes,alpha=0.1)

y$cumhaz \%>\%
  ggplot(aes(x=exp(-Lambda)))+
  stat_ecdf()+
  geom_abline(slope=1)
}
\seealso{
Other Genealogy processes: 
\code{\link{moran}},
\code{\link{sirws}}
}
\concept{Genealogy processes}
