% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbdp.R
\name{lbdp}
\alias{lbdp}
\alias{LBDP}
\alias{runLBDP}
\alias{continueLBDP}
\alias{lbdp_exact}
\alias{lbdp_pomp}
\title{Linear birth-death-sampling model}
\usage{
runLBDP(time, t0 = 0, lambda = 2, mu = 1, psi = 1, frac = 0, n0 = 5)

continueLBDP(object, time, lambda = NA, mu = NA, psi = NA, frac = NA)

lbdp_exact(data, lambda, mu, psi, n0 = 1)

lbdp_pomp(
  data,
  lambda,
  mu,
  psi,
  n0 = 1,
  t0 = 0,
  method = c("gillespie", "euler"),
  delta.t = NULL
)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{lambda}{per capita birth rate}

\item{mu}{per capita recovery rate.}

\item{psi}{per capita sampling rate.}

\item{frac}{fraction of batch sampling}

\item{n0}{initial population size}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}

\item{data}{data frame containing the genealogy event times and event codes.}

\item{method}{integration method}

\item{delta.t}{Euler step-size when \code{method="euler"} is chosen}
}
\value{
An object of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{LBDP}.

\code{lbdp_exact} returns the log likelihood of the genealogy.
Note that the time since the most recent sample is informative.
}
\description{
The genealogy process induced by a simple linear birth-death process
with constant-rate sampling.
}
\details{
\code{lbdp_exact} gives the exact likelihood of a linear birth-death process, conditioned on \eqn{n_0 = 0}{n0=0} (Stadler, 2010, Thm 3.5).
The derivation is also given in comments in the code.

The \code{data} argument should in the format returned by \code{\link{newick2df}}.

\code{lbdp_pomp} constructs a \pkg{pomp} object containing a given set of data and a linear birth-death-sampling process.

It is assumed that \code{data} is in the format returned by \code{\link{newick2df}}.
}
\examples{
simulate("LBDP",time=4) |> plot(points=TRUE)

simulate("LBDP",lambda=2,mu=1,psi=3,n0=1,time=1) |>
  simulate(time=10,lambda=1) |>
  plot()

simulate("LBDP",time=4) |>
  lineages() |>
  plot()

}
\references{
\Stadler2010
}
\seealso{
More example genealogy processes: 
\code{\link{batch}()},
\code{\link{lbdpwr}},
\code{\link{moranwr}},
\code{\link{moran}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sirwr}},
\code{\link{sir}}
}
\concept{Genealogy processes}
