% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages.R
\name{lineages}
\alias{lineages}
\title{Lineage-count function}
\usage{
lineages(object, prune = TRUE)
}
\arguments{
\item{object}{\code{gpsim} object.}

\item{prune}{logical; prune the genealogy?}
}
\value{
A tibble containing the lineage count function.
}
\description{
The number of lineages through time
}
