% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineages.R
\name{lineages}
\alias{lineages}
\alias{plot.gplin}
\title{Lineage-count function}
\usage{
lineages(object, prune = TRUE, obscure = TRUE)

\method{plot}{gplin}(x, ..., palette = scales::hue_pal(l = 30, h = c(220, 580)))
}
\arguments{
\item{object}{\code{gpsim} object.}

\item{prune}{logical; prune the genealogy?}

\item{obscure}{logical; obscure the demes?}

\item{x}{object of class \sQuote{gpsim}}

\item{...}{passed to \code{\link{treeplot}}}

\item{palette}{color palette for branches.
This can be furnished either as a function or a vector of colors.
If this is a function, it should take a single integer argument, the number of colors required.
If it is a vector, ....}
}
\value{
A \code{\link[tibble]{tibble}} containing the lineage count function.
If the genealogy has been obscured (the default), the number in the \code{lineages}
column is the total number of lineages present at the times in the \code{time} column.
If the genealogy has not been obscured (\code{obscure = FALSE}), the deme-specific
lineage counts are returned.

The \code{\link[tibble]{tibble}} returned by \code{lineages} has a \code{\link[=plot.gplin]{plot}} method.
}
\description{
The number of lineages through time
}
\examples{
library(tidyverse)

pal <- c("#00274c","#ffcb05")

simulate("SIIR",time=3) -> x
plot_grid(
  x |> plot(),
  x |> lineages() |> plot.gplin(),
  x |> plot(obscure=FALSE,palette=pal),
  x |> lineages(obscure=FALSE) |>
    getElement(1) |>
    plot(palette=pal,legend.position=c(0.8,0.9)),
  align="v",axis="b",
  ncol=2,byrow=FALSE
)
}
