% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{moran}
\alias{moran}
\alias{Moran}
\alias{runMoran}
\alias{continueMoran}
\title{Moran genealogy process}
\usage{
runMoran(time, t0 = 0, n = 100, mu = 1, psi = 1)

continueMoran(object, time, mu = NA, psi = NA)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{n}{population size}

\item{mu}{event rate}

\item{psi}{sampling rate.}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}

\item{stationary}{logical;
if TRUE, the process will be initialized in its stationary distribution.}
}
\value{
An object of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{"Moran"}.
}
\description{
Run the simulator.
}
\seealso{
Other Genealogy processes: 
\code{\link{lbdp}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sir}}
}
\concept{Genealogy processes}
