% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{moran}
\alias{moran}
\alias{Moran}
\alias{runMoran}
\alias{continueMoran}
\alias{moran_exact}
\title{The classical Moran model}
\usage{
runMoran(time, t0 = 0, n = 100, mu = 1, psi = 1)

continueMoran(object, time, mu = NA, psi = NA)

moran_exact(x, n = 100, mu = 1, psi = 1)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{n}{population size}

\item{mu}{event rate}

\item{psi}{sampling rate.}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}

\item{x}{genealogy in \pkg{phylopomp} format (i.e., an object that inherits from \sQuote{gpgen}).}
}
\value{
\code{runMoran} and \code{continueMoran} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{Moran}.

\code{moran_exact} returns the log likelihood of the genealogy.
}
\description{
The Markov genealogy process induced by the classical Moran process,
in which birth/death events occur at a constant rate and the
population size remains constant.
}
\details{
\code{moran_exact} gives the exact log likelihood of a genealogy under the uniformly-sampled Moran process.
}
\seealso{
More example genealogy processes: 
\code{\link{lbdp}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sir}}
}
\concept{Genealogy processes}
