% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{moran}
\alias{moran}
\alias{MGP}
\alias{mgp}
\alias{playMoran}
\title{Moran genealogy process.}
\usage{
playMoran(
  data = NULL,
  ...,
  n,
  mu = n,
  t0 = 0,
  times,
  tree = FALSE,
  ill = FALSE,
  stationary = TRUE
)
}
\arguments{
\item{data}{optional data frame;
output from \code{playMoran}.}

\item{...}{additional arguments; ignored.}

\item{n}{population size}

\item{mu}{Moran event rate}

\item{t0}{initial time}

\item{times}{times at which output is requested.}

\item{tree}{logical; represent the genealogical tree in Newick format?}

\item{ill}{logical;
return an illustration?}

\item{stationary}{logical;
should the initial genealogy be drawn from the stationary distribution?}
}
\value{
A \code{tibble} with \code{state} attribute.
}
\description{
Run the MGP simulator.
}
\examples{
playMoran(n=5,mu=5,times=0:10,t0=0,tree=TRUE,ill=TRUE) -> x
playMoran(x,times=11:20,tree=TRUE) -> x
plot(x)

playMoran(n=20,mu=20,times=0:20,stationary=FALSE,tree=TRUE,ill=TRUE) -> x
plot(x)

y <- getInfo(x,prune=FALSE)
plot(y,points=TRUE)

library(dplyr)
playMoran(n=5,mu=10,times=c(0,seq(100,200,by=25)),stationary=FALSE,ill=TRUE) \%>\%
  mutate(grob=diagram(illustration)) -> x

}
\seealso{
Other Genealogy processes: 
\code{\link{lbdp}},
\code{\link{sirws}}
}
\concept{Genealogy processes}
