% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{moran}
\alias{moran}
\alias{MGP}
\alias{mgp}
\alias{playMoran}
\alias{getInfo.Moran_gpsim}
\title{Moran genealogy process.}
\usage{
playMoran(
  data = NULL,
  ...,
  n,
  mu,
  t0 = 0,
  times,
  tree = FALSE,
  stationary = TRUE
)

\method{getInfo}{Moran_gpsim}(data, ..., prune = TRUE, tree = TRUE)
}
\arguments{
\item{data}{optional data frame;
output from \code{playMoran}.}

\item{...}{additional arguments; ignored.}

\item{n}{population size}

\item{mu}{Moran event rate}

\item{t0}{initial time}

\item{times}{times at which output is requested.}

\item{tree}{logical;
represent the genealogical tree in Newick format?}

\item{stationary}{logical;
should the initial genealogy be drawn from the stationary distribution?}

\item{prune}{Prune the tree?}
}
\value{
A \code{tibble} with \code{state} attribute.
}
\description{
Run the MGP simulator.
}
\examples{
playMoran(n=5,mu=5,times=1:10,t0=0,tree=TRUE) -> x
playMoran(x,times=11:20,tree=TRUE) -> x
plot(x)

playMoran(n=20,mu=20,times=1:20,stationary=FALSE,tree=TRUE) -> x
plot(x)

y <- getInfo(x,prune=FALSE)
plot(y,points=TRUE)
}
\seealso{
Other Genealogy processes: 
\code{\link{sirws}}
}
\concept{Genealogy processes}
