% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R, R/moran_exact.R
\name{moran}
\alias{moran}
\alias{Moran}
\alias{runMoran}
\alias{continueMoran}
\alias{moran_exact}
\title{The classical Moran model}
\usage{
runMoran(time, t0 = 0, mu = 1, psi = 1, n = 100)

continueMoran(object, time, mu = NA, psi = NA)

moran_exact(x, n = 100, mu = 1, psi = 1)
}
\arguments{
\item{time}{end timepoint of simulation}

\item{t0}{beginning timepoint of simulation}

\item{mu}{per capita event rate}

\item{psi}{per capita sampling rate}

\item{n}{population size}

\item{object}{a previously computed simulation}

\item{x}{genealogy in \pkg{phylopomp} format (i.e., an object that inherits from \sQuote{gpgen}).}
}
\value{
\code{runMoran} and \code{continueMoran} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{Moran}.

\code{moran_exact} returns the log likelihood of the genealogy.
}
\description{
The Markov genealogy process induced by the classical Moran process, in which birth/death events occur at a constant rate and the population size remains constant.
}
\details{
\code{moran_exact} gives the exact log likelihood of a genealogy under the uniformly-sampled Moran process.
}
\references{
\Moran1958
}
\seealso{
More example genealogy processes:
\code{\link{bdei}},
\code{\link{bdss}},
\code{\link{lbdp}},
\code{\link{mers}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sir}},
\code{\link{strains}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
