% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newick.R
\name{newick}
\alias{newick}
\title{Newick output}
\usage{
newick(object, prune = TRUE, obscure = TRUE)
}
\arguments{
\item{object}{\code{gpsim} object.}

\item{prune}{logical; prune the genealogy?}

\item{obscure}{logical; obscure the demes?}
}
\value{
A string in Newick format.
}
\description{
Extract a Newick-format description of a genealogy.
}
\examples{
simulate("SIIR",time=1) |> newick()

}
