% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newick2df.R
\name{newick2df}
\alias{newick2df}
\title{Tree in newick to dataframe}
\usage{
newick2df(tree)
}
\arguments{
\item{tree}{tree data in newick format.}
}
\value{
A data frame could be used as \code{pomp} input.
}
\description{
Covert a tree in newick to a data frame, containing event times, number of lineages, sample indicator, direct descent indicator, and coalescent indicator.
}
\examples{
playSIRwS(beta=2,gamma=1,psi=2,S0=100,I0=2,times=5,t0=0,tree=TRUE) -> x
y <- getInfo(x)
newick2df(y$tree) -> z
head(z)
}
