% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_newick}
\alias{parse_newick}
\title{parse a Newick string}
\usage{
parse_newick(x, t0 = 0, tf = NA)
}
\arguments{
\item{x}{character; the Newick description.
See Details for specifics.}

\item{t0}{numeric; the root time.}

\item{tf}{numeric; the current or final time.}
}
\value{
An object of class \dQuote{gpgen}.
}
\description{
Parses a Newick description and returns a binary version of the genealogy.
}
