% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreach.R, R/getinfo.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{foreach}
\alias{\%dopar\%}
\alias{registerDoSEQ}
\alias{as.yaml}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{foreach}{\code{\link[foreach:foreach]{\%dopar\%}}, \code{\link[foreach]{foreach}}, \code{\link[foreach]{registerDoSEQ}}}

  \item{yaml}{\code{\link[yaml]{as.yaml}}}
}}

