% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagram.R, R/foreach.R, R/pomp.R, R/yaml.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{viewport}
\alias{plot_grid}
\alias{foreach}
\alias{\%dopar\%}
\alias{registerDoSEQ}
\alias{bake}
\alias{stew}
\alias{freeze}
\alias{as.yaml}
\alias{read_yaml}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{cowplot}{\code{\link[cowplot]{plot_grid}}}

  \item{foreach}{\code{\link[foreach:foreach]{\%dopar\%}}, \code{\link[foreach]{foreach}}, \code{\link[foreach]{registerDoSEQ}}}

  \item{grid}{\code{\link[grid]{viewport}}}

  \item{pomp}{\code{\link[pomp]{bake}}, \code{\link[pomp:bake]{freeze}}, \code{\link[pomp:bake]{stew}}}

  \item{yaml}{\code{\link[yaml]{as.yaml}}, \code{\link[yaml]{read_yaml}}}
}}

