% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2i2r2.R
\name{s2i2r2}
\alias{s2i2r2}
\alias{S2I2R2}
\alias{runS2I2R2}
\alias{continueS2I2R2}
\title{Two-host infection model with waning, immigration, and demography.}
\usage{
runS2I2R2(
  time,
  t0 = 0,
  Beta11 = 4,
  Beta12 = 0,
  Beta22 = 4,
  gamma1 = 1,
  gamma2 = 1,
  psi1 = 1,
  psi2 = 0,
  omega1 = 0,
  omega2 = 0,
  b1 = 0,
  b2 = 0,
  d1 = 0,
  d2 = 0,
  iota1 = 0,
  iota2 = 0,
  S1_0 = 100,
  S2_0 = 100,
  I1_0 = 0,
  I2_0 = 10,
  R1_0 = 0,
  R2_0 = 0
)

continueS2I2R2(
  object,
  time,
  Beta11 = NA,
  Beta12 = NA,
  Beta22 = NA,
  gamma1 = NA,
  gamma2 = NA,
  psi1 = NA,
  psi2 = NA,
  omega1 = NA,
  omega2 = NA,
  b1 = NA,
  b2 = NA,
  d1 = NA,
  d2 = NA,
  iota1 = NA,
  iota2 = NA
)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{Beta11, Beta22}{transmission rates within species 1 and 2, respectively}

\item{Beta12}{transmission from species 2 to species 1}

\item{gamma1, gamma2}{recovery rates for species 1 and 2, respectively}

\item{psi1, psi2}{per capita sampling rates}

\item{omega1, omega2}{rates of waning of immunity}

\item{b1, b2}{per capita birth rates}

\item{d1, d2}{per capita death rates}

\item{iota1, iota2}{infection importation rates}

\item{S1_0, S2_0}{initial sizes of susceptible populations}

\item{I1_0, I2_0}{initial sizes of infected populations}

\item{R1_0, R2_0}{initial sizes of immune populations}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}
}
\value{
\code{runS2I2R2} and \code{continueS2I2R2} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{S2I2R2}.
}
\description{
The population is structured by infection progression and host species.
}
\seealso{
More example genealogy processes:
\code{\link{lbdp}},
\code{\link{moran}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sir}},
\code{\link{twospecies}}
}
\concept{Genealogy processes}
