% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.R, R/seir_pomp.R
\name{seir}
\alias{seir}
\alias{SEIR}
\alias{runSEIR}
\alias{runSEIRS}
\alias{continueSEIR}
\alias{continueSEIRS}
\alias{seirs_pomp}
\title{Classical susceptible-exposed-infected-recovered model}
\usage{
runSEIR(
  time,
  t0 = 0,
  Beta = 4,
  sigma = 1,
  gamma = 1,
  psi = 1,
  chi = 0,
  omega = 0,
  pop = 100,
  S0 = 0.9,
  E0 = 0.05,
  I0 = 0.05,
  R0 = 0
)

runSEIRS(
  time,
  t0 = 0,
  Beta = 4,
  sigma = 1,
  gamma = 1,
  psi = 1,
  chi = 0,
  omega = 0,
  pop = 100,
  S0 = 0.9,
  E0 = 0.05,
  I0 = 0.05,
  R0 = 0
)

continueSEIR(
  object,
  time,
  Beta = NA,
  sigma = NA,
  gamma = NA,
  psi = NA,
  chi = NA,
  omega = NA
)

continueSEIRS(
  object,
  time,
  Beta = NA,
  sigma = NA,
  gamma = NA,
  psi = NA,
  chi = NA,
  omega = NA
)

seirs_pomp(x, Beta, sigma, gamma, psi, chi = 0, omega = 0, S0, E0, I0, R0, pop)
}
\arguments{
\item{time}{end timepoint of simulation}

\item{t0}{beginning timepoint of simulation}

\item{Beta}{transmission rate}

\item{sigma}{progression rate}

\item{gamma}{recovery rate}

\item{psi}{per capita non-destructive sampling rate}

\item{chi}{per capita destructive sampling rate}

\item{omega}{rate of waning of immunity}

\item{pop}{host population size}

\item{S0}{initial fraction of population susceptible to infection}

\item{E0}{initial fraction of population in exposed class}

\item{I0}{initial fraction of population in infectious class}

\item{R0}{initial fraction of population immune to infection}

\item{object}{a previously computed simulation}

\item{x}{genealogy in \pkg{phylopomp} format.}
}
\value{
\code{runSEIR} and \code{continueSEIR} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{SEIR}.

\code{seirs_pomp} returns a \sQuote{pomp} object.
}
\description{
The population is structured by infection progression.
There is an incubation period that must pass before an infected
host becomes infectious. The duration of this period is \eqn{1/\sigma}.
The infectious period is \eqn{1/\gamma}.
}
\details{
\code{seirs_pomp} constructs a \sQuote{pomp} object containing a given set of data and an SEIRS model.
}
\examples{
simulate("SEIR",Beta=2,sigma=2,gamma=1,psi=2,S0=1,I0=0.01,time=5) |>
  simulate(Beta=5,gamma=2,time=10,psi=3) |>
  plot()

runSEIR(Beta=3,gamma=1,psi=2,S0=1,I0=0.2,R0=0,time=5,t0=-1) |>
  plot(points=TRUE,obscure=FALSE)

runSEIR(Beta=3,gamma=0.1,psi=0.2,S0=1,I0=0.05,R0=0,time=2,t0=0) -> x
plot_grid(plotlist=list(plot(x,points=TRUE),diagram(x)),
  ncol=1,rel_heights=c(4,1))

simulate("SEIR",sigma=1,omega=1,time=20,I0=0.04) |> plot(obscure=FALSE)

simulate("SEIR",sigma=1,omega=1,time=20,I0=0.04) |>
  lineages(obscure=FALSE) |>
  plot()
}
\references{
\King2024
}
\seealso{
More example genealogy processes:
\code{\link{bdei}},
\code{\link{bdss}},
\code{\link{lbdp}},
\code{\link{mers}},
\code{\link{moran}},
\code{\link{mtbd2}},
\code{\link{mtbd3}},
\code{\link{s2i2r2}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link[=simulate]{simulate()}},
\code{\link{sir}},
\code{\link{strains}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
