% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir.R
\name{seir}
\alias{seir}
\alias{SEIR}
\alias{runSEIR}
\alias{continueSEIR}
\alias{seirs_pomp}
\title{Classical susceptible-exposed-infected-recovered model}
\usage{
runSEIR(
  time,
  t0 = 0,
  Beta = 4,
  sigma = 1,
  gamma = 1,
  psi = 1,
  delta = 0,
  S0 = 100,
  E0 = 5,
  I0 = 5,
  R0 = 0
)

continueSEIR(
  object,
  time,
  Beta = NA,
  sigma = NA,
  gamma = NA,
  psi = NA,
  delta = NA
)

seirs_pomp(x, Beta, sigma, gamma, psi, delta = 0, S0, E0, I0, R0)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{Beta}{transmission rate}

\item{sigma}{progression rate}

\item{gamma}{recovery rate}

\item{psi}{per capita sampling rate}

\item{delta}{rate of waning of immunity}

\item{S0}{initial size of susceptible population}

\item{E0}{initial size of exposed population}

\item{I0}{initial size of infected population}

\item{R0}{initial size of immune population}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}

\item{x}{genealogy in \pkg{phylopomp} format.}
}
\value{
\code{runSEIR} and \code{continueSEIR} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{SEIR}.
}
\description{
The population is structured by infection progression.
}
\details{
\code{seirs_pomp} constructs a \pkg{pomp} object containing a given set of data and an SEIRS model.
}
\examples{
simulate("SEIR",Beta=2,sigma=2,gamma=1,psi=2,S0=1000,I0=5,time=5) |>
  simulate(Beta=5,gamma=2,time=10,psi=3) |>
  plot()

runSEIR(Beta=3,gamma=1,psi=2,S0=20,I0=5,R0=0,time=5,t0=-1) |>
  plot(points=TRUE,obscure=FALSE)

runSEIR(Beta=3,gamma=0.1,psi=0.2,S0=100,I0=5,R0=0,time=2,t0=0) -> x
plot_grid(plotlist=list(plot(x,points=TRUE),diagram(x)),
  ncol=1,rel_heights=c(4,1))

simulate("SEIR",sigma=1,delta=1,time=20,I0=4) |> plot(obscure=FALSE)

simulate("SEIR",sigma=1,delta=1,time=20,I0=4) |>
  lineages(obscure=FALSE) |>
  plot()
}
\seealso{
More example genealogy processes: 
\code{\link{lbdp}},
\code{\link{moran}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sir}}
}
\concept{Genealogy processes}
