% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si2r.R
\name{si2r}
\alias{si2r}
\alias{SI2R}
\alias{runSI2R}
\alias{continueSI2R}
\title{Two-deme model of superspreading}
\usage{
runSI2R(
  time,
  t0 = 0,
  Beta = 5,
  kappa = 2,
  gamma = 1,
  omega = 0,
  psi = 1,
  etaL = 1,
  etaH = 3,
  pop = 500,
  S0 = 0.98,
  IL0 = 0.02,
  IH0 = 0,
  R0 = 0
)

continueSI2R(
  object,
  time,
  Beta = NA,
  kappa = NA,
  gamma = NA,
  omega = NA,
  psi = NA,
  etaL = NA,
  etaH = NA
)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{Beta}{transmission rate}

\item{kappa}{super-spreading transmission factor}

\item{gamma}{recovery rate}

\item{omega}{rate of waning of immunity}

\item{psi}{sampling rate}

\item{etaL}{rate of transition from low-spreading to super-spreading behavior}

\item{etaH}{rate of transition from super-spreading to low-spreading behavior}

\item{pop}{population size}

\item{S0}{initial size of susceptible population}

\item{IL0}{initial size of low-spreading population}

\item{IH0}{initial size of super-spreading population}

\item{R0}{initial size of immune population}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}
}
\value{
\code{runSI2R} and \code{continueSI2R} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{SI2R}.
}
\description{
Deme L consists of "low-rate spreaders" that transmit at rate \code{Beta}.
Deme H consists of "superspreaders" who transmit at a higher rate
\code{kappa*Beta}.
}
\seealso{
More example genealogy processes:
\code{\link{lbdp}},
\code{\link{moran}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{siir}},
\code{\link[=simulate]{simulate()}},
\code{\link{sir}},
\code{\link{strains}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
