% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siir.R
\name{siir}
\alias{siir}
\alias{runSIIR}
\alias{SIIR}
\title{SIIR with sampling simulator.}
\usage{
runSIIR(
  Beta1 = 2,
  Beta2 = 2,
  gamma = 1,
  psi = 1,
  S0 = 100,
  I1_0 = 1,
  I2_0 = 1,
  R0 = 0,
  t0 = 0,
  time = 1
)
}
\arguments{
\item{Beta1, Beta2}{transmission rates from each of the infectious classes.}

\item{gamma}{recovery rate.}

\item{psi}{sampling rate.}

\item{S0}{initial size of susceptible population.}

\item{I1_0}{initial size of I2 population.}

\item{I2_0}{initial size of I2 population.}

\item{R0}{initial size of recovered population.}

\item{t0}{initial time}

\item{time}{end timepoint of simulation.}
}
\value{
An object of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{"SIR"}.
}
\description{
Run the simulator.
}
\examples{
runSIIR(Beta1=2,Beta2=4,gamma=1,psi=5,S0=200,I1_0=5,I2_0=5,R0=0,time=3,t0=0) |>
  continue(Beta1=5,Beta2=10,gamma=2,time=10,psi=10) |>
  plot()

runSIIR(Beta1=10,Beta2=8,gamma=1,psi=2,S0=300,I1_0=2,I2_0=20,R0=0,time=0,t0=-1) |>
  continue(psi=10,time=2) |>
  plot(points=TRUE,prune=FALSE)

library(ggplot2)
runSIIR(Beta1=10,Beta2=8,gamma=1,psi=2,S0=300,I1_0=2,I2_0=20,R0=0,time=5) |>
  lineages() |>
  ggplot(aes(x=time,y=lineages))+
  geom_step()
}
\seealso{
Other Genealogy processes: 
\code{\link{continue}()},
\code{\link{sir}}
}
\concept{Genealogy processes}
