% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siir.R
\name{siir}
\alias{siir}
\alias{playSIIR}
\alias{SIIR}
\title{SIIR with sampling simulator.}
\usage{
playSIIR(
  data = NULL,
  Beta1,
  Beta2,
  gamma,
  psi,
  S0,
  I1_0,
  I2_0,
  R0,
  t0 = 0,
  times,
  tree = FALSE,
  ill = FALSE
)
}
\arguments{
\item{data}{optional data frame; output from \code{playSIR}.}

\item{Beta1, Beta2}{transmission rates from each of the infectious classes.}

\item{gamma}{recovery rate.}

\item{psi}{sampling rate.}

\item{S0}{initial size of susceptible population.}

\item{I1_0}{initial size of I2 population.}

\item{I2_0}{initial size of I2 population.}

\item{R0}{initial size of recovered population.}
}
\value{
A \code{tibble} with \code{state} attribute.
}
\description{
Run the simulator.
}
\examples{
playSIR(Beta=2,gamma=1,psi=2,S0=1000,I0=5,R0=0,times=0:5,t0=0,tree=TRUE) -> x
playSIR(x,times=6:10,psi=1,tree=TRUE) -> x
plot(x)

playSIR(Beta=3,gamma=1,psi=2,S0=10,I0=5,R0=0,times=0:5,t0=-1,tree=TRUE) -> x
plot(x,points=TRUE)

y <- getInfo(x)
plot(y,points=TRUE)

library(ggplot2)
y$lineages |>
  ggplot(aes(x=time,y=lineages))+
  geom_step()
}
\seealso{
Other Genealogy processes: 
\code{\link{sir}}
}
\concept{Genealogy processes}
