% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siir.R
\name{siir}
\alias{siir}
\alias{playSIIR}
\alias{SIIR}
\title{SIIR with sampling simulator.}
\usage{
playSIIR(
  data = NULL,
  Beta1,
  Beta2,
  gamma,
  psi,
  S0,
  I1_0,
  I2_0,
  R0,
  t0 = 0,
  times,
  tree = FALSE,
  compact = TRUE
)
}
\arguments{
\item{data}{data frame containing the genealogy in the format returned by \code{\link{newick2df}}.}

\item{Beta1, Beta2}{transmission rates from each of the infectious classes.}

\item{gamma}{recovery rate.}

\item{psi}{sampling rate.}

\item{S0}{initial size of susceptible population.}

\item{I1_0}{initial size of I2 population.}

\item{I2_0}{initial size of I2 population.}

\item{R0}{initial size of recovered population.}

\item{t0}{initial time}

\item{times}{times at which output is requested.}

\item{tree}{logical; represent the genealogical tree in Newick format?}

\item{compact}{logical; return the tree in compact format?}
}
\value{
A \code{tibble} with \code{state} attribute.
}
\description{
Run the simulator.
}
\examples{
playSIIR(Beta1=2,Beta2=4,gamma=1,psi=5,S0=100,I1_0=5,I2_0=1,R0=0,times=0:5,t0=0,tree=TRUE) -> x
playSIIR(x,times=6:10,psi=5,tree=TRUE) -> x
plot(x)

playSIIR(Beta1=3,Beta2=4,gamma=1,psi=2,S0=10,I1_0=5,I2_0=1,R0=0,times=0:5,t0=-1,tree=TRUE) -> x
plot(x,points=TRUE)

y <- getInfo(x)
plot(y,points=TRUE)
}
\seealso{
Other Genealogy processes: 
\code{\link{sir}}
}
\concept{Genealogy processes}
