% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siir.R
\name{siir}
\alias{siir}
\alias{SIIR}
\alias{runSIIR}
\alias{continueSIIR}
\title{Two-strain SIR model.}
\usage{
runSIIR(
  time,
  t0 = 0,
  Beta1 = 5,
  Beta2 = 5,
  gamma = 1,
  psi1 = 1,
  psi2 = 0,
  sigma12 = 0,
  sigma21 = 0,
  delta = 0,
  frac = 0,
  S0 = 500,
  I1_0 = 10,
  I2_0 = 10,
  R0 = 0
)

continueSIIR(
  object,
  time,
  Beta1 = NA,
  Beta2 = NA,
  gamma = NA,
  psi1 = NA,
  psi2 = NA,
  sigma12 = NA,
  sigma21 = NA,
  delta = NA,
  frac = NA
)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{Beta1, Beta2}{transmission rates from each of the infectious classes.}

\item{gamma}{recovery rate.}

\item{psi1, psi2}{sampling rates.}

\item{sigma12, sigma21}{movement rates from deme 1 to 2 and 2 to 1, respectively}

\item{delta}{rate of loss of immunity}

\item{frac}{fraction of batch sampling}

\item{S0}{initial size of susceptible population.}

\item{I1_0}{initial size of I2 population.}

\item{I2_0}{initial size of I2 population.}

\item{R0}{initial size of recovered population.}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}
}
\value{
An object of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{SIIR}.
}
\description{
Two distinct pathogen strains compete for susceptibles.
}
\examples{
simulate("SIIR",time=3,psi1=1,psi2=0) |>
  simulate(Beta1=2,gamma=2,time=10,psi1=10,psi2=1) |>
  plot()

runSIIR(Beta1=10,Beta2=8,
  S0=200,I1_0=10,I2_0=8,R0=0,time=0,t0=-1) |>
  simulate(psi1=10,time=2) |>
  plot(points=TRUE,obscure=FALSE)

simulate("SIIR",Beta1=2,Beta2=50,gamma=1,psi1=2,
  S0=300,I1_0=20,I2_0=2,time=5) |>
  lineages() |>
  plot.gplin()

simulate("SIIR",Beta1=2,Beta2=50,gamma=1,psi1=0,frac=.5,
         S0=300,I1_0=20,I2_0=2,time=5) |>
  batch() |>
  plot()
}
\seealso{
More example genealogy processes: 
\code{\link{batch}()},
\code{\link{lbdpwr}},
\code{\link{lbdp}},
\code{\link{moranwr}},
\code{\link{moran}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{simulate}()},
\code{\link{sirwr}},
\code{\link{sir}}
}
\concept{Genealogy processes}
