% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\alias{continue}
\alias{continue.gpsim}
\title{simulate}
\usage{
simulate(model, ...)

continue(object, ...)

\method{continue}{gpsim}(object, time, ...)
}
\arguments{
\item{model}{name of model to simulate}

\item{...}{additional arguments to the model-specific simulation functions}

\item{object}{previously computed \sQuote{gpsim} object}

\item{time}{end timepoint of simulation}
}
\value{
An object of \sQuote{gpsim} class.
}
\description{
Simulate Markov genealogy processes
}
\section{Note}{

One cannot change initial conditions or \code{t0} with \code{continue}.
}

\seealso{
Other Genealogy processes: 
\code{\link{siir}},
\code{\link{sir}}
}
\concept{Genealogy processes}
