% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\alias{simulate.default}
\alias{simulate.character}
\alias{simulate.gpsim}
\title{simulate}
\usage{
simulate(object, ...)

\method{simulate}{default}(object, ...)

\method{simulate}{character}(object, time, ...)

\method{simulate}{gpsim}(object, time, ...)
}
\arguments{
\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}

\item{...}{additional arguments to the model-specific simulation functions}

\item{time}{end timepoint of simulation}
}
\value{
An object of \sQuote{gpsim} class.
}
\description{
Simulate Markov genealogy processes
}
\details{
When \code{object} is of class \sQuote{gpsim}, i.e., the result of a genealogy-process
simulation, \code{simulate} acts to continue the simulation to a later timepoint.
Note that, one cannot change initial conditions or \code{t0} when continuing a simulation.
}
\references{
\King2024

\King2022
}
\seealso{
More example genealogy processes:
\code{\link{lbdp}},
\code{\link{moran}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{sir}},
\code{\link{strains}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
