% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir}
\alias{sir}
\alias{playSIR}
\alias{SIR}
\alias{sir_pomp}
\title{SIR with sampling simulator.}
\usage{
playSIR(
  data = NULL,
  Beta,
  gamma,
  psi,
  S0,
  I0,
  R0,
  t0 = 0,
  times,
  tree = FALSE,
  compact = TRUE
)

sir_pomp(data, Beta, gamma, psi, S0, I0, R0, t0 = 0)
}
\arguments{
\item{data}{data frame containing the genealogy in the format returned by \code{\link{newick2df}}.}

\item{Beta}{transmission rate.}

\item{gamma}{recovery rate.}

\item{psi}{sampling rate.}

\item{S0}{initial size of susceptible population.}

\item{I0}{initial size of infected population.}

\item{R0}{initial size of recovered population.}

\item{t0}{initial time}

\item{times}{times at which output is requested.}

\item{tree}{logical; represent the genealogical tree in Newick format?}

\item{compact}{logical; return the tree in compact format?}
}
\value{
A \code{tibble} with \code{state} attribute.
}
\description{
Run the simulator.
}
\details{
\code{sir_pomp} constructs a \pkg{pomp} object containing a given set of data and a SIR model.
}
\examples{
playSIR(Beta=2,gamma=1,psi=2,S0=1000,I0=5,R0=0,times=0:5,t0=0,tree=TRUE) -> x
playSIR(x,Beta=5,gamma=2,times=6:10,psi=3,tree=TRUE) -> x
plot(x)

playSIR(Beta=3,gamma=1,psi=2,S0=10,I0=5,R0=0,times=0:5,t0=-1,tree=TRUE) -> x
plot(x,points=TRUE)

y <- getInfo(x)
plot(y,points=TRUE)

playSIR(Beta=30,gamma=1,psi=2,S0=100,I0=5,R0=0,times=0.1,t0=0) -> x
y <- getInfo(x,prune=FALSE,compact=FALSE)
plot(y,points=TRUE)
}
\seealso{
Other Genealogy processes: 
\code{\link{siir}}
}
\concept{Genealogy processes}
