% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R, R/sir_pomp.R
\name{sir}
\alias{sir}
\alias{SIR}
\alias{SIRS}
\alias{runSIR}
\alias{runSIRS}
\alias{continueSIR}
\alias{continueSIRS}
\alias{sir_pomp}
\alias{sirs_pomp}
\title{Classical susceptible-infected-recovered model}
\usage{
runSIR(
  time,
  t0 = 0,
  Beta = 2,
  gamma = 1,
  psi = 1,
  omega = 0,
  S0 = 100,
  I0 = 2,
  R0 = 0
)

runSIRS(
  time,
  t0 = 0,
  Beta = 2,
  gamma = 1,
  psi = 1,
  omega = 0,
  S0 = 100,
  I0 = 2,
  R0 = 0
)

continueSIR(object, time, Beta = NA, gamma = NA, psi = NA, omega = NA)

runSIRS(
  time,
  t0 = 0,
  Beta = 2,
  gamma = 1,
  psi = 1,
  omega = 0,
  S0 = 100,
  I0 = 2,
  R0 = 0
)

continueSIRS(object, time, Beta = NA, gamma = NA, psi = NA, omega = NA)

sir_pomp(x, Beta, gamma, psi, omega = 0, S0, I0, R0, t0 = 0)

sirs_pomp(x, Beta, gamma, psi, omega = 0, S0, I0, R0, t0 = 0)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{Beta}{transmission rate.}

\item{gamma}{recovery rate.}

\item{psi}{sampling rate.}

\item{omega}{immunity waning rate}

\item{S0, I0, R0}{initial sizes of susceptible, infected, and recovered populations, respectively.}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}

\item{x}{genealogy in \pkg{phylopomp} format (i.e., an object that inherits from \sQuote{gpgen}).}
}
\value{
\code{runSIR} and \code{continueSIR} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{SIR}.

\code{sir_pomp} and \code{sirs_pomp} return \sQuote{pomp} objects.
}
\description{
A single, unstructured population of hosts.
}
\details{
\code{sir_pomp} constructs a \sQuote{pomp} object containing a given set of data and a SIR model.
}
\examples{
simulate("SIR",Beta=2,gamma=1,psi=2,S0=1000,I0=5,time=5) |>
  simulate(Beta=5,gamma=2,time=10,psi=3) |>
  plot()

runSIR(Beta=3,gamma=1,psi=2,S0=20,I0=5,R0=0,time=5,t0=-1) |>
  plot(points=TRUE)

runSIR(Beta=3,gamma=0.1,psi=0.2,S0=100,I0=5,R0=0,time=2,t0=0) -> x
plot_grid(plotlist=list(plot(x,points=TRUE),diagram(x)),
  ncol=1,rel_heights=c(4,1))

simulate("SIRS",omega=1,time=20,I0=4) |> plot()
simulate("SIRS",omega=1,time=20,I0=4) |> lineages() |> plot()
}
\references{
\King2024

\King2022
}
\seealso{
More example genealogy processes: 
\code{\link{lbdp}},
\code{\link{moran}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{twospecies}}
}
\concept{Genealogy processes}
