% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir}
\alias{sir}
\alias{runSIR}
\alias{SIR}
\alias{sir_pomp}
\title{SIR with sampling simulator.}
\usage{
runSIR(
  Beta = 2,
  gamma = 1,
  psi = 1,
  S0 = 100,
  I0 = 2,
  R0 = 0,
  t0 = 0,
  time = 1
)

sir_pomp(data, Beta, gamma, psi, S0, I0, R0, t0 = 0)
}
\arguments{
\item{Beta}{transmission rate.}

\item{gamma}{recovery rate.}

\item{psi}{sampling rate.}

\item{S0}{initial size of susceptible population.}

\item{I0}{initial size of infected population.}

\item{R0}{initial size of recovered population.}

\item{t0}{initial time}

\item{time}{end timepoint of simulation.}

\item{data}{data frame containing the genealogy in the format returned by \code{\link{newick2df}}.}
}
\value{
An object of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{"SIR"}.
}
\description{
Run the simulator.
}
\details{
\code{sir_pomp} constructs a \pkg{pomp} object containing a given set of data and a SIR model.
}
\examples{
runSIR(Beta=2,gamma=1,psi=2,S0=1000,I0=5,R0=0,time=5) |>
  continue(Beta=5,gamma=2,time=10,psi=3) |>
  plot()

runSIR(Beta=3,gamma=1,psi=2,S0=20,I0=5,R0=0,time=5,t0=-1) |>
  plot(points=TRUE)

library(cowplot)
runSIR(Beta=3,gamma=0.1,psi=0.2,S0=100,I0=5,R0=0,time=2,t0=0) -> x
plot_grid(plotlist=list(plot(x,points=TRUE),diagram(x)),
  ncol=1,rel_heights=c(4,1))
}
\seealso{
Other Genealogy processes: 
\code{\link{continue}()},
\code{\link{siir}}
}
\concept{Genealogy processes}
