% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir}
\alias{sir}
\alias{SIR}
\alias{runSIR}
\alias{continueSIR}
\alias{sir_pomp}
\title{Classical susceptible-infected-recovered model}
\usage{
runSIR(
  time,
  t0 = 0,
  Beta = 2,
  gamma = 1,
  psi = 1,
  delta = 0,
  S0 = 100,
  I0 = 2,
  R0 = 0
)

continueSIR(object, time, Beta = NA, gamma = NA, psi = NA, delta = NA)

sir_pomp(data, Beta, gamma, psi, delta = 0, S0, I0, R0, t0 = 0)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{Beta}{transmission rate.}

\item{gamma}{recovery rate.}

\item{psi}{sampling rate.}

\item{delta}{immunity waning rate}

\item{S0}{initial size of susceptible population.}

\item{I0}{initial size of infected population.}

\item{R0}{initial size of recovered population.}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}

\item{data}{data frame containing the genealogy in the format returned by \code{\link{newick2df}}.}
}
\value{
An object of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{SIR}.
}
\description{
A single, unstructured population of hosts.
}
\details{
\code{sir_pomp} constructs a \pkg{pomp} object containing a given set of data and a SIR model.
}
\examples{
simulate("SIR",Beta=2,gamma=1,psi=2,S0=1000,I0=5,time=5) |>
  simulate(Beta=5,gamma=2,time=10,psi=3) |>
  plot()

runSIR(Beta=3,gamma=1,psi=2,S0=20,I0=5,R0=0,time=5,t0=-1) |>
  plot(points=TRUE)

runSIR(Beta=3,gamma=0.1,psi=0.2,S0=100,I0=5,R0=0,time=2,t0=0) -> x
plot_grid(plotlist=list(plot(x,points=TRUE),diagram(x)),
  ncol=1,rel_heights=c(4,1))

simulate("SIR",delta=1,time=20,I0=4) |> plot()
simulate("SIR",delta=1,time=20,I0=4) |> lineages() |> plot()
}
\seealso{
More example genealogy processes: 
\code{\link{lbdp}},
\code{\link{moran}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()}
}
\concept{Genealogy processes}
