% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirwr.R
\name{sirwr}
\alias{sirwr}
\alias{SIRwr}
\alias{runSIRwr}
\alias{continueSIRwr}
\title{SIR model with S, M, L segments and reassortment}
\usage{
runSIRwr(
  time,
  t0 = 0,
  Beta = 2,
  gamma = 1,
  psi = 1,
  delta = 0,
  frac = 0,
  rhoS = 0,
  rhoM = 0.1,
  rhoL = 0,
  rhoSM = 0,
  rhoSL = 0,
  rhoML = 0,
  S0 = 100,
  I0 = 2,
  R0 = 0
)

continueSIRwr(
  object,
  time,
  Beta = NA,
  gamma = NA,
  psi = NA,
  delta = NA,
  frac = NA,
  rhoS = NA,
  rhoM = NA,
  rhoL = NA,
  rhoSM = NA,
  rhoSL = NA,
  rhoML = NA
)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{Beta}{transmission rate.}

\item{gamma}{recovery rate.}

\item{psi}{sampling rate.}

\item{delta}{immunity waning rate}

\item{frac}{fraction of batch sampling}

\item{rhoS}{reassortment rate of S segment}

\item{rhoM}{reassortment rate of M segment}

\item{rhoL}{reassortment rate of L segment}

\item{rhoSM}{reassortment rate of both S and M segment}

\item{rhoSL}{reassortment rate of both S and L segment}

\item{rhoML}{reassortment rate of both M and L segment}

\item{S0}{initial size of susceptible population.}

\item{I0}{initial size of infected population.}

\item{R0}{initial size of recovered population.}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}
}
\value{
An object of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{SIRwr}.
}
\description{
A single, unstructured population of hosts.
}
\examples{
simulate("SIRwr",Beta=2,gamma=1,psi=2,frac=0,
         rhoS=.5, rhoM=.1, rhoL=.3,
         rhoSM=0, rhoSL=.2, rhoML = 0,
         S0=100,I0=10,time=5) |>
  plot(prune=TRUE, compact=TRUE, points=TRUE)

set.seed(28494712L)
runSIRwr(Beta=3,gamma=1,psi=2,frac=0,
         rhoS=0,rhoM=.5,rhoL=0,
         rhoSM=0, rhoSL=0, rhoML = 0,
         S0=20,I0=5,R0=0,time=5,t0=-1) |>
  plot(prune=TRUE, compact=TRUE, points=TRUE, t0=-1)

set.seed(284914712L)
runSIRwr(Beta=3,gamma=1,psi=0,frac=.8,
         rhoS=0,rhoM=.5,rhoL=0,
         rhoSM=0, rhoSL=0, rhoML = 0,
         S0=100,I0=10,R0=0,time=1,t0=0) |>
  batch() |>
  plot()
}
\seealso{
More example genealogy processes: 
\code{\link{batch}()},
\code{\link{lbdpwr}},
\code{\link{lbdp}},
\code{\link{moranwr}},
\code{\link{moran}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sir}}
}
\concept{Genealogy processes}
