% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strains.R, R/strains_pomp.R
\name{strains}
\alias{strains}
\alias{Strains}
\alias{runStrains}
\alias{continueStrains}
\alias{strains_pomp}
\title{Three strains compete for a single susceptible pool.}
\usage{
runStrains(
  time,
  t0 = 0,
  Beta1 = 5,
  Beta2 = 5,
  Beta3 = 5,
  gamma = 1,
  chi1 = 1,
  chi2 = 0,
  chi3 = 0,
  pop = 1e+05,
  S_0 = 0.8,
  I1_0 = 0.01,
  I2_0 = 0.01,
  I3_0 = 0.01,
  R_0 = 0.2
)

continueStrains(
  object,
  time,
  Beta1 = NA,
  Beta2 = NA,
  Beta3 = NA,
  gamma = NA,
  chi1 = NA,
  chi2 = NA,
  chi3 = NA
)

strains_pomp(
  x,
  Beta1,
  Beta2,
  Beta3,
  gamma,
  chi1,
  chi2,
  chi3,
  pop,
  S_0,
  I1_0,
  I2_0,
  I3_0,
  R_0
)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{Beta1}{transmission rate for strain 1}

\item{Beta2}{transmission rate for strain 2}

\item{Beta3}{transmission rate for strain 3}

\item{gamma}{recovery rate}

\item{chi1}{(destructive) sampling rate for strain 1}

\item{chi2}{(destructive) sampling rate for strain 2}

\item{chi3}{(destructive) sampling rate for strain 3}

\item{pop}{population size}

\item{S_0}{initial susceptible fraction}

\item{I1_0}{initial fraction of population infected by strain 1}

\item{I2_0}{initial fraction of population infected by strain 2}

\item{I3_0}{initial fraction of population infected by strain 2}

\item{R_0}{initial fraction of population immune}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}

\item{x}{genealogy in \pkg{phylopomp} format (i.e., an object that inherits from \sQuote{gpgen}).}
}
\value{
\code{runStrains} and \code{continueStrains} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{Strains}.

\code{strains_pomp} returns a \sQuote{pomp} object.
}
\description{
The three demes are three distinct pathogen strains that
compete for susceptibles. Sampling is assumed destructive,
i.e., sampled individuals are removed from the infectious pool.
}
\details{
\code{strains_pomp} constructs a \sQuote{pomp} object containing a given set of data and the Strains model.
}
\seealso{
More example genealogy processes:
\code{\link{lbdp}},
\code{\link{mers}},
\code{\link{moran}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link[=simulate]{simulate()}},
\code{\link{sir}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
