% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strains.R, R/strains_pomp.R
\name{strains}
\alias{strains}
\alias{Strains}
\alias{runStrains}
\alias{continueStrains}
\alias{strains_pomp}
\title{Three strains compete for a single susceptible pool.}
\usage{
runStrains(
  time,
  t0 = 0,
  Beta1 = 5,
  Beta2 = 5,
  Beta3 = 5,
  gamma = 1,
  psi1 = 1,
  psi2 = 0,
  psi3 = 0,
  S0 = 1,
  I1_0 = 0.01,
  I2_0 = 0.01,
  I3_0 = 0.01,
  R0 = 0,
  pop = 10000
)

continueStrains(
  object,
  time,
  Beta1 = NA,
  Beta2 = NA,
  Beta3 = NA,
  gamma = NA,
  psi1 = NA,
  psi2 = NA,
  psi3 = NA
)

strains_pomp(
  x,
  Beta1,
  Beta2,
  Beta3,
  gamma,
  psi1,
  psi2,
  psi3,
  pop,
  S0,
  I1_0,
  I2_0,
  I3_0,
  R0
)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{Beta1, Beta2, Beta3}{transmission rate for strains 1, 2, 3, respectively}

\item{gamma}{recovery rate}

\item{psi1, psi2, psi3}{sampling rates}

\item{S0, I1_0, I2_0, I3_0, R0}{initial conditions; non-negative numbers that specify the relative occupancies of the compartments at the inital time.}

\item{pop}{host population size}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}

\item{x}{genealogy in \pkg{phylopomp} format (i.e., an object that inherits from \sQuote{gpgen}).}
}
\value{
\code{runStrains} and \code{continueStrains} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{Strains}.

\code{strains_pomp} returns a \sQuote{pomp} object.
}
\description{
The three demes are three distinct pathogen strains that compete for susceptibles.
}
\details{
\code{strains_pomp} constructs a \sQuote{pomp} object containing a given set of data and the Strains model.
}
\seealso{
More example genealogy processes:
\code{\link{lbdp}},
\code{\link{moran}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sir}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
