% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strains.R
\name{strains}
\alias{strains}
\alias{Strains}
\alias{runStrains}
\alias{continueStrains}
\title{Three strains compete for a single susceptible pool.}
\usage{
runStrains(
  time,
  t0 = 0,
  Beta1 = 5,
  Beta2 = 5,
  Beta3 = 5,
  gamma = 1,
  omega = 0,
  psi1 = 1,
  psi2 = 0,
  psi3 = 0,
  S0 = 10000,
  I1_0 = 10,
  I2_0 = 10,
  I3_0 = 10,
  R0 = 0
)

continueStrains(
  object,
  time,
  Beta1 = NA,
  Beta2 = NA,
  Beta3 = NA,
  gamma = NA,
  omega = NA,
  psi1 = NA,
  psi2 = NA,
  psi3 = NA
)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{Beta1, Beta2, Beta3}{transmission rate for strains 1, 2, 3, respectively}

\item{gamma}{recovery rate}

\item{omega}{rate of waning of immunity}

\item{psi1, psi2, psi3}{sampling rates}

\item{S0}{initial size of susceptible population}

\item{I1_0, I2_0, I3_0}{initial numbers of strain-specific infections}

\item{R0}{initial size of immune population}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}
}
\value{
\code{runStrains} and \code{continueStrains} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{Strains}.
}
\description{
The three demes are three distinct pathogen strains that compete for susceptibles.
}
\seealso{
More example genealogy processes:
\code{\link{lbdp}},
\code{\link{moran}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sir}},
\code{\link{twospecies}},
\code{\link{twoundead}}
}
\concept{Genealogy processes}
