% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplot.R
\name{treeplot}
\alias{treeplot}
\title{Fancy tree plotter}
\usage{
treeplot(data, ladderize = TRUE, points = FALSE)
}
\arguments{
\item{data}{Output of one of the \code{play} functions.}

\item{ladderize}{Ladderize?}

\item{points}{Show nodes and tips?}
}
\value{
A printable \code{ggtree} object.
}
\description{
Plots a genealogical tree.
}
