% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeplot.R
\name{treeplot}
\alias{treeplot}
\title{Fancy tree plotter}
\usage{
treeplot(
  tree,
  time = NULL,
  root_time = 0,
  ladderize = TRUE,
  points = FALSE,
  palette = "Set1"
)
}
\arguments{
\item{tree}{character; tree representation in Newick format.}

\item{time}{numeric; times of the genealogies.}

\item{root_time}{numeric; time of the root.}

\item{ladderize}{Ladderize?}

\item{points}{Show nodes and tips?}

\item{palette}{character; color palette to use for branches.
See \code{\link[ggplot2]{scale_color_brewer}} for details.}
}
\value{
A printable \code{ggplot} object.
}
\description{
Plots a genealogical tree.
}
