% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twospecies.R
\name{twospecies}
\alias{twospecies}
\alias{TwoSpecies}
\alias{runTwoSpecies}
\alias{continueTwoSpecies}
\title{Two-host infection model with waning, immigration, demography, and spillover.}
\usage{
runTwoSpecies(
  time,
  t0 = 0,
  Beta11 = 4,
  Beta12 = 0,
  Beta21 = 0,
  Beta22 = 4,
  gamma1 = 1,
  gamma2 = 1,
  psi1 = 1,
  psi2 = 0,
  omega1 = 0,
  omega2 = 0,
  b1 = 0,
  b2 = 0,
  d1 = 0,
  d2 = 0,
  iota1 = 0,
  iota2 = 0,
  S1_0 = 100,
  S2_0 = 100,
  I1_0 = 0,
  I2_0 = 10,
  R1_0 = 0,
  R2_0 = 0
)

continueTwoSpecies(
  object,
  time,
  Beta11 = NA,
  Beta12 = NA,
  Beta21 = NA,
  Beta22 = NA,
  gamma1 = NA,
  gamma2 = NA,
  psi1 = NA,
  psi2 = NA,
  omega1 = NA,
  omega2 = NA,
  b1 = NA,
  b2 = NA,
  d1 = NA,
  d2 = NA,
  iota1 = NA,
  iota2 = NA
)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{Beta11, Beta22}{transmission rates within species 1 and species 2, respectively}

\item{Beta12}{transmission from species 2 to species 1}

\item{Beta21}{transmission from species 1 to species 2}

\item{gamma1, gamma2}{recovery rates}

\item{psi1, psi2}{per capita sampling rates}

\item{omega1, omega2}{rates of waning of immunity}

\item{b1, b2}{per capita birth rates}

\item{d1, d2}{per capita death rates}

\item{iota1, iota2}{imported infection rates}

\item{S1_0, S2_0}{initial sizes of the susceptible populations in each species}

\item{I1_0, I2_0}{initial sizes of infected populations}

\item{R1_0, R2_0}{initial sizes of immune populations}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}
}
\value{
\code{runTwoSpecies} and \code{continueTwoSpecies} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{TwoSpecies}.
}
\description{
The population is structured by infection progression and host species.
}
\seealso{
More example genealogy processes: 
\code{\link{lbdp}},
\code{\link{moran}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sir}}
}
\concept{Genealogy processes}
