% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoundead.R
\name{twoundead}
\alias{twoundead}
\alias{TwoUndead}
\alias{runTwoUndead}
\alias{continueTwoUndead}
\title{Two-host infection model with waning, immigration, demography, and spillover. Hosts are culled upon sampling with a given probability. This is identical to the TwoSpecies model with the exception that dead lineages are not pruned. Instead, they become *ghosts*.}
\usage{
runTwoUndead(
  time,
  t0 = 0,
  Beta11 = 4,
  Beta12 = 0,
  Beta21 = 0,
  Beta22 = 4,
  gamma1 = 1,
  gamma2 = 1,
  psi1 = 1,
  psi2 = 0,
  c1 = 1,
  c2 = 1,
  omega1 = 0,
  omega2 = 0,
  b1 = 0,
  b2 = 0,
  d1 = 0,
  d2 = 0,
  iota1 = 0,
  iota2 = 0,
  S1_0 = 100,
  S2_0 = 100,
  I1_0 = 0,
  I2_0 = 10,
  R1_0 = 0,
  R2_0 = 0
)

continueTwoUndead(
  object,
  time,
  Beta11 = NA,
  Beta12 = NA,
  Beta21 = NA,
  Beta22 = NA,
  gamma1 = NA,
  gamma2 = NA,
  psi1 = NA,
  psi2 = NA,
  c1 = NA,
  c2 = NA,
  omega1 = NA,
  omega2 = NA,
  b1 = NA,
  b2 = NA,
  d1 = NA,
  d2 = NA,
  iota1 = NA,
  iota2 = NA
)
}
\arguments{
\item{time}{final time}

\item{t0}{initial time}

\item{Beta11}{transmission rate within species 1}

\item{Beta12}{transmission from species 2 to species 1}

\item{Beta21}{transmission from species 1 to species 2}

\item{Beta22}{transmission rate within species 2}

\item{gamma1}{species 1 recovery rate}

\item{gamma2}{species 2 recovery rate}

\item{psi1}{per capita sampling rate for species 1}

\item{psi2}{per capita sampling rate for species 2}

\item{c1}{probability that a sampled (positive) host of species 1 is culled}

\item{c2}{probability that a sampled (positive) host of species 2 is culled}

\item{omega1}{rate of waning of immunity for species 1}

\item{omega2}{rate of waning of immunity for species 2}

\item{b1}{per capita birth rate for species 1}

\item{b2}{per capita birth rate for species 2}

\item{d1}{per capita death rate for species 1}

\item{d2}{per capita death rate for species 2}

\item{iota1}{imported infections for species 1}

\item{iota2}{imported infections for species 2}

\item{S1_0}{initial size of species 1 susceptible population}

\item{S2_0}{initial size of species 2 susceptible population}

\item{I1_0}{initial size of species 1 infected population}

\item{I2_0}{initial size of species 2 infected population}

\item{R1_0}{initial size of species 1 immune population}

\item{R2_0}{initial size of species 2 immune population}

\item{object}{either the name of the model to simulate
\emph{or} a previously computed \sQuote{gpsim} object}
}
\value{
\code{runTwoUndead} and \code{continueTwoUndead} return objects of class \sQuote{gpsim} with \sQuote{model} attribute \dQuote{TwoUndead}.
}
\description{
The population is structured by infection progression and host species.
}
\seealso{
More example genealogy processes:
\code{\link{lbdp}},
\code{\link{moran}},
\code{\link{s2i2r2}},
\code{\link{seir}},
\code{\link{si2r}},
\code{\link{siir}},
\code{\link{simulate}()},
\code{\link{sir}},
\code{\link{strains}},
\code{\link{twospecies}}
}
\concept{Genealogy processes}
