% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{yaml}
\alias{yaml}
\title{YAML output}
\usage{
yaml(object, prune = TRUE, obscure = TRUE, trace = FALSE)
}
\arguments{
\item{object}{\code{gpsim} object.}

\item{prune}{logical; prune the genealogy?}

\item{obscure}{logical; obscure the demes?}

\item{trace}{logical; trace the lineages?}
}
\value{
A string in yaml format.
}
\description{
Human- and machine-readable description
}
\examples{
simulate("SIIR",time=1) |> yaml() |> cat()

}
